perf(i18n): 拆 cloud-console 字典到独立 chunk，app 入口 -50% gzipped
surface-text-dictionaries.ts 原本 3750 行 / 2335 条 cloud-console 专属字典
通过 @yinjie/i18n/index 的 \`export *\` 被 catalog-loaders.ts 静态拉到所有
surface 的 entry chunk —— app web 端被迫下载 cloud-console 的所有字符串。

改造为 registry 模式：
- surface-text-dictionaries.ts (slim, ~40 行)：只剩 registry API
  (registerSurfaceTextDictionaryProvider + 同步 getSurfaceTextDictionary lookup)
- surface-text-dictionaries-cloud-console.ts (3750 行)：原有数据 + 在文件末尾
  side-effect 调 registerSurfaceTextDictionaryProvider 把自己挂上来
- packages/i18n/package.json exports 加 ./runtime/surface-text-dictionaries-
  cloud-console，cloud-console main.tsx + 测试文件用 side-effect import
  触发注册
- @yinjie/i18n/index.ts 不再 re-export 重数据：app/admin/wiki tree-shake 干净

构建对比 (entry index-*.js gzipped)：
  88KB → 44KB （–50%）
  整个 app critical path gzipped ~257KB → ~213KB（–17%）

接口兼容：getSurfaceTextDictionary(surface, locale) 同步签名不变，cloud-
console-i18n.ts 老调用点零改动；只需要 cloud-console main.tsx + 测试加一行
side-effect import 触发注册。

verify:
- pnpm --filter @yinjie/i18n typecheck ✓
- pnpm --filter @yinjie/app typecheck / build:fast ✓
- pnpm --filter @yinjie/cloud-console typecheck ✓
- 公网 http://1gw06751dd053.vicp.fun:29490/ 仍返回 app HTML ✓
