+++ packages/i18n/src/runtime/surface-text-dictionaries.ts
@@ -1,3750 +1,38 @@
 import type { I18nAppSurface, SupportedLocale } from "../locales";
 
-type LocaleDictionary = Partial<Record<SupportedLocale, Record<string, string>>>;
+// Surface 级 text dictionary registry（运行时注册）。
+//
+// 历史背景：cloud-console 有 ~3750 行字典数据，原本直接放在这个文件里通过
+// 静态 export，被 catalog-loaders.ts → app-locale-provider.tsx → @yinjie/i18n
+// 的 entry re-export 拖到所有 surface 的 critical bundle，移动 web 端凭空多
+// 下 ~70KB raw / ~20KB gzipped。
+//
+// 现在改成 registry 模式：surface 自己有字典数据时，单独导入对应 data 文件
+// （比如 ./surface-text-dictionaries-cloud-console），那个 data 文件会主动
+// 调 registerSurfaceTextDictionaryProvider 把自己挂上来。其它 surface 不
+// 导入对应 data 文件 → 整坨数据 tree-shake 走。
+//
+// 兼容性：getSurfaceTextDictionary 的同步函数签名不变，cloud-console-i18n.ts
+// 等老调用点不需要改；只需 cloud-console main.tsx 加一行 side-effect import。
 
-const cloudConsoleText: LocaleDictionary = {
-  "zh-CN": {
-    "Yinjie Cloud Ops": "隐界云运维",
-    "Cloud World Console": "云世界控制台",
-    "Manage phone-based world provisioning, wake old worlds, inspect lifecycle jobs, and track instance health from one place.":
-      "集中管理基于手机号的建世界申请、唤醒旧世界、检查生命周期任务，并追踪实例健康状态。",
-    "Enter CLOUD_ADMIN_SECRET": "输入 CLOUD_ADMIN_SECRET",
-    "Admin access required": "需要管理密钥",
-    "Enter CLOUD_ADMIN_SECRET to unlock the console. Cloud requests are paused until a secret is saved locally.":
-      "请输入 CLOUD_ADMIN_SECRET 以解锁控制台。保存到本地之前，云平台请求会暂停。",
-    "CLOUD_ADMIN_SECRET is required.": "请先输入 CLOUD_ADMIN_SECRET。",
-    "CLOUD_ADMIN_SECRET is invalid.": "CLOUD_ADMIN_SECRET 无效。",
-    "Cloud admin session is invalid or expired.":
-      "云平台管理会话无效或已过期。",
-    "Admin secret saved locally. Short-lived admin tokens will refresh automatically.":
-      "管理密钥已保存在本地，短期管理令牌会自动刷新。",
-    "Admin secret cleared.": "管理密钥已清除。",
-    "Admin secret saved locally. Console uses short-lived admin tokens.":
-      "管理密钥已保存在本地，控制台将使用短期管理令牌。",
-    "Admin secret is missing.": "缺少管理密钥。",
-    "Admin secret configured": "管理密钥已配置",
-    "Admin secret missing": "管理密钥未配置",
-    "Cloud access": "云平台访问",
-    "Cloud operations": "云平台运维",
-    Navigation: "导航",
-    More: "更多",
-    Setup: "设置",
-    "Request detail": "申请详情",
-    "Availability and drift": "可用性与漂移",
-    "Applications and handoffs": "申请与交接",
-    "Instances and health": "实例与健康",
-    "Queue and leases": "队列与租约",
-    "Access audit": "访问审计",
-    "Durable tasks": "持久化任务",
-    "World request detail": "世界申请详情",
-    "Review one phone-based world application and its delivery state.":
-      "复核单个手机号建世界申请及其交付状态。",
-    "Review application status, projected world state, and manual delivery handoffs.":
-      "复核申请状态、预计世界状态和人工交付交接。",
-    "World detail": "世界详情",
-    "Inspect instance placement, runtime status, bootstrap material, and lifecycle jobs.":
-      "检查实例部署、运行状态、启动材料和生命周期任务。",
-    "Track provider placement, power state, heartbeat freshness, and operator attention.":
-      "追踪供应方部署、电源状态、心跳新鲜度和运维关注项。",
-    "Session audit": "会话审计",
-    "Review admin access, risk groups, revocation paths, and source activity.":
-      "复核管理访问、风险分组、吊销路径和来源活动。",
-    "Durable recovery": "持久化恢复",
-    "Replay exhausted compensation tasks and clear stale failures from one queue.":
-      "在一个队列中重放耗尽重试的补偿任务并清除陈旧失败。",
-    "Monitor world availability, lifecycle jobs, request flow, and cloud runtime drift.":
-      "监控世界可用性、生命周期任务、申请流转和云运行时漂移。",
-    "Request id": "请求 ID",
+type DictionaryProvider = (locale: SupportedLocale) => Map<string, string>;
 
-    Dashboard: "仪表盘",
-    Requests: "申请",
-    Worlds: "世界",
-    Jobs: "任务",
-    Sessions: "会话",
-    "Waiting Sync": "等待同步",
-    Users: "用户",
-    Plans: "套餐",
-    Configs: "配置",
-    "Invite Audit": "邀请审计",
-    "Invite audit": "邀请审计",
-    "Revenue Sharing": "收益分成",
-    "SaaS operations": "SaaS 运维",
-    "Cloud monetization": "云平台收益",
-    "User detail": "用户详情",
-    "Fleet Dashboard": "舰队仪表盘",
-    "World requests": "世界申请",
-    "Managed worlds": "托管世界",
-    "Lifecycle jobs": "生命周期任务",
-    "Admin sessions": "管理会话",
-    "Waiting session sync": "等待会话同步",
-    "Request guidance": "申请指引",
-    "Lifecycle summary": "生命周期摘要",
-    "Bootstrap package": "启动包",
-    "Attention Queue": "关注队列",
-    "Batch actions": "批量操作",
+const dictionaryProviders = new Map<I18nAppSurface, DictionaryProvider>();
 
-    "Ready worlds": "就绪世界",
-    "Critical attention": "严重关注",
-    "Recovery queued": "恢复已排队",
-    "Provider drift": "Provider 漂移",
-    "Instance footprint": "实例占用",
-    "Request alerts": "申请告警",
-    "Active jobs": "活跃任务",
-    "Failed jobs": "失败任务",
-    "Superseded jobs": "已替代任务",
-    "Running jobs": "运行中任务",
-    "Lease expired jobs": "租约过期任务",
-    "Delayed jobs": "延迟任务",
-    "Pending approvals": "待审批",
-    "Provisioning handoffs": "建世界交接",
-    "Rejected requests": "已拒绝申请",
-    "Disabled requests": "已停用申请",
-    "Recent requests": "最近申请",
-    "Recent lifecycle jobs": "最近生命周期任务",
-    "Recent failed jobs": "最近失败任务",
-    "Recent superseded jobs": "最近已替代任务",
-    "No active jobs right now.": "当前没有活跃任务。",
-    "No failed jobs right now.": "当前没有失败任务。",
-    "No superseded jobs right now.": "当前没有已替代任务。",
-    "No attention items right now.": "当前没有关注项。",
-    "No recent requests.": "没有最近申请。",
-
-    All: "全部",
-    Unknown: "未知",
-    "Not available": "不可用",
-    None: "无",
-    "Not set": "未设置",
-    Save: "保存",
-    Edit: "编辑",
-    Copy: "复制",
-    Open: "打开",
-    Close: "关闭",
-    Cancel: "取消",
-    Confirm: "确认",
-    Clear: "清除",
-    Reset: "重置",
-    Retry: "重试",
-    Refresh: "刷新",
-    Apply: "应用",
-    Download: "下载",
-    Export: "导出",
-    Search: "搜索",
-    Submit: "提交",
-    Update: "更新",
-    Approve: "通过",
-    Reject: "拒绝",
-    Disable: "停用",
-    Enable: "启用",
-    Resume: "恢复",
-    Suspend: "挂起",
-    Reconcile: "对账",
-    Rotate: "轮换",
-    Revoke: "吊销",
-    Hide: "隐藏",
-    Show: "显示",
-    "Show more": "显示更多",
-    "Show less": "收起",
-    "View all": "查看全部",
-    "View jobs": "查看任务",
-    "View world": "查看世界",
-    "View worlds": "查看世界",
-    "View requests": "查看申请",
-    "View sessions": "查看会话",
-    "Open dashboard": "打开仪表盘",
-    "Open requests": "打开申请",
-    "Open worlds": "打开世界",
-    "Open jobs": "打开任务",
-    "Open sessions": "打开会话",
-    "Open waiting sync": "打开等待同步",
-    "Open sessions permalink": "打开会话固定链接",
-    "Open requests permalink": "打开申请固定链接",
-    "Open worlds permalink": "打开世界固定链接",
-    "Open jobs permalink": "打开任务固定链接",
-    "Open waiting-sync permalink": "打开等待同步固定链接",
-
-    Status: "状态",
-    Phone: "手机号",
-    "Phone number": "手机号",
-    World: "世界",
-    Request: "申请",
-    Job: "任务",
-    Session: "会话",
-    "World id": "世界 ID",
-    "Job id": "任务 ID",
-    "Session id": "会话 ID",
-    Name: "名称",
-    Provider: "Provider",
-    Region: "区域",
-    Zone: "可用区",
-    Strategy: "策略",
-    Health: "健康",
-    "Desired state": "期望状态",
-    "Power state": "电源状态",
-    Attention: "关注",
-    Severity: "严重级别",
-    Reason: "原因",
-    Message: "消息",
-    Error: "错误",
-    "Last error": "最近错误",
-    "Failure code": "失败代码",
-    "Failure message": "失败消息",
-    "API base URL": "API 基础 URL",
-    "Admin URL": "管理后台 URL",
-    "Instance URL": "实例 URL",
-    "Callback token": "回调令牌",
-    "Runtime status": "运行时状态",
-    "Bootstrap config": "启动配置",
-    "Docker compose": "Docker Compose",
-    "Runtime env overlay": "运行时环境覆盖",
-    "Callback endpoints": "回调端点",
-    "Lifecycle actions": "生命周期操作",
-    "Job queue": "任务队列",
-    "Queue state": "队列状态",
-    "Task type": "任务类型",
-    "Task key": "任务 Key",
-    Target: "目标",
-    Context: "上下文",
-    Attempt: "尝试",
-    Available: "可用",
-    Updated: "更新于",
-    Created: "创建于",
-    Finished: "完成于",
-    Started: "开始于",
-    "Last accessed": "最近访问",
-    "Last interactive": "最近交互",
-    "Last booted": "最近启动",
-    "Last heartbeat": "最近心跳",
-    "Last suspended": "最近挂起",
-    "Last operation": "最近操作",
-    "Observed at": "观测时间",
-    "Lease owner": "租约持有者",
-    "Lease expires": "租约到期",
-    "Latest update": "最近更新",
-    "Latest event": "最近事件",
-    "Last generated": "最近生成",
-
-    Ready: "就绪",
-    Pending: "待处理",
-    Provisioning: "建世界中",
-    Active: "活跃",
-    Rejected: "已拒绝",
-    Disabled: "已停用",
-    Failed: "失败",
-    Succeeded: "成功",
-    Running: "运行中",
-    Queued: "已排队",
-    Sleeping: "休眠",
-    Stopping: "停止中",
-    Starting: "启动中",
-    Stopped: "已停止",
-    Expired: "已过期",
-    Revoked: "已吊销",
-    Current: "当前",
-    Healthy: "健康",
-    Critical: "严重",
-    Warning: "警告",
-    Normal: "正常",
-    Watch: "观察",
-    "All statuses": "全部状态",
-    "All reasons": "全部原因",
-    "All sessions": "全部会话",
-    "Current only": "仅当前",
-    Ascending: "升序",
-    Descending: "降序",
-    "Last used": "最近使用",
-    "Revoked at": "吊销时间",
-    Logout: "退出登录",
-    "Manual revocation": "手动吊销",
-    "Refresh token reuse": "刷新令牌复用",
-    "Multiple active sessions": "多个活跃会话",
-    "Repeated revocations": "重复吊销",
-
-    "Request search": "申请搜索",
-    "World search": "世界搜索",
-    "Waiting sync search": "等待同步搜索",
-    "Projected world status": "预计世界状态",
-    "Projected desired state": "预计期望状态",
-    "Waiting sync status": "等待同步状态",
-    "Waiting sync task type": "等待同步任务类型",
-    "Waiting sync page size": "等待同步分页大小",
-    "Session id, IP, client, revoker": "会话 ID、IP、客户端、吊销方",
-    "Search world, phone, job, lease...": "搜索世界、手机号、任务、租约...",
-    "task key, target, context, or error": "任务 key、目标、上下文或错误",
-    "request id, phone, world name, or status detail":
-      "申请 ID、手机号、世界名或状态详情",
-    "world id, phone, name, provider, or endpoint":
-      "世界 ID、手机号、名称、供应方或端点",
-    "world id, phone, email, name, provider, or endpoint":
-      "世界 ID、手机号、邮箱、名称、供应方或端点",
-
-    "Copy callback token": "复制回调令牌",
-    "Copy runtime env overlay": "复制运行时环境覆盖",
-    "Copy docker compose snippet": "复制 Docker Compose 片段",
-    "Context task review": "上下文任务复核",
-    "Recent task receipts": "最近任务回执",
-    "Unable to load waiting sync tasks": "无法加载等待同步任务",
-    "Unable to load requests": "无法加载申请",
-    "Unable to load worlds": "无法加载世界",
-    "Unable to load jobs": "无法加载任务",
-    "Unable to load admin sessions": "无法加载管理会话",
-
-    "Refresh world": "刷新世界",
-    "Refresh phone": "刷新手机号",
-    "Invalidate phone": "使手机号失效",
-    "Replay failed": "重放失败项",
-    "Clear failed": "清除失败项",
-    "Replay task": "重放任务",
-    "Clear task": "清除任务",
-    "Batch replay failed": "批量重放失败项",
-    "Batch clear failed": "批量清除失败项",
-    "Filtered replay failed": "按筛选重放失败项",
-    "Filtered clear failed": "按筛选清除失败项",
-    "Switch status to All or Failed before running batch failed-task actions.":
-      "执行批量失败任务操作前，请将状态切换为全部或失败。",
-    "All failed tasks across every page.": "所有页面中的全部失败任务。",
-    "Focus snapshot appears when the current query exactly matches a visible context or target.":
-      "当前查询精确匹配可见上下文或目标时，会显示聚焦快照。",
-    "Add a context or target query to export a tighter investigation snapshot.":
-      "添加上下文或目标查询，以导出更聚焦的排查快照。",
-
-    "Rotate the callback token?": "要轮换回调令牌吗？",
-    "Revoke admin session?": "要吊销管理会话吗？",
-    "Revoke selected admin sessions?": "要吊销选中的管理会话吗？",
-    "Revoke all matching admin sessions?": "要吊销所有匹配的管理会话吗？",
-    "Revoke matching risk groups?": "要吊销匹配的风险分组吗？",
-    "Revoke source group?": "要吊销来源分组吗？",
-    "Close dialog": "关闭弹窗",
-
-    "queue: all": "队列：全部",
-    "queue: running": "队列：运行中",
-    "queue: lease expired": "队列：租约过期",
-    "queue: delayed": "队列：延迟",
-    "queue: pending": "队列：待处理",
-    "sort by: updated": "排序：更新时间",
-    "sort by: created": "排序：创建时间",
-    "sort by: status": "排序：状态",
-    "direction: descending": "方向：降序",
-    "direction: ascending": "方向：升序",
-
-    // SaaS / Revenue Sharing pages
-    Account: "账号",
-    Subscription: "订阅",
-    Expires: "到期时间",
-    Inviter: "邀请人",
-    Plan: "套餐",
-    Previous: "上一页",
-    Next: "下一页",
-    Inactive: "未启用",
-    "Not loaded": "未加载",
-    Policy: "策略",
-    Version: "版本",
-    Payable: "应付",
-    Held: "暂扣",
-    Enabled: "启用",
-    Currency: "货币",
-    "External ref": "外部引用",
-    Payees: "收益人",
-    Payee: "收益人",
-    Participant: "参与方",
-    Amount: "金额",
-    Unassigned: "未分配",
-    Ledger: "账本",
-    Events: "事件",
-    Preview: "预览",
-    "Usage type": "用量类型",
-    Character: "角色",
-    Gross: "毛收入",
-    Occurred: "发生时间",
-    "Contribution type": "贡献类型",
-    Contributor: "贡献者",
-    State: "状态",
-    reversed: "已撤销",
-    active: "活跃",
-    banned: "已封禁",
-    archived: "已归档",
-    expired: "已过期",
-    none: "无",
-    rewarded: "已奖励",
-    rejected: "已拒绝",
-    trial: "试用",
-    purchasable: "可购买",
-    "All account states": "全部账号状态",
-    "All subscription states": "全部订阅状态",
-    "All redemption states": "全部兑换状态",
-    "Search phone": "搜索手机号",
-    "Search phone or code": "搜索手机号或邀请码",
-    "Loading SaaS users...": "正在加载 SaaS 用户...",
-    "Failed to load users.": "用户加载失败。",
-    "No cloud users matched the current filters.": "当前筛选条件下没有匹配的云用户。",
-    "Loading cloud user...": "正在加载云用户...",
-    "Failed to load cloud user.": "云用户加载失败。",
-    "SaaS user": "SaaS 用户",
-    "Account:": "账号：",
-    "Subscription:": "订阅：",
-    "Current plan:": "当前套餐：",
-    "Expires at:": "到期时间：",
-    "Invite code:": "邀请码：",
-    "World status:": "世界状态：",
-    "World:": "世界：",
-    "Open world": "打开世界",
-    "Back to users": "返回用户列表",
-    "Manual grant": "手动赠送",
-    "Grant days": "赠送天数",
-    "Account state": "账号状态",
-    "Ban reason": "封禁原因",
-    Ban: "封禁",
-    Unban: "解封",
-    "Subscription history": "订阅记录",
-    "No subscription records found.": "暂无订阅记录。",
-    "Invite history": "邀请记录",
-    "No invite rewards recorded.": "暂无邀请奖励记录。",
-    "Loading subscription plans...": "正在加载订阅套餐...",
-    "Failed to load subscription plans.": "订阅套餐加载失败。",
-    "Save plan": "保存套餐",
-    "New plan": "新建套餐",
-    "Subscription plan saved.": "订阅套餐已保存。",
-    code: "代码",
-    name: "名称",
-    durationDays: "时长（天）",
-    priceCents: "价格（分）",
-    currency: "货币",
-    sortOrder: "排序",
-    description: "描述",
-    "Loading cloud configs...": "正在加载云配置...",
-    "Failed to load cloud configs.": "云配置加载失败。",
-    "Config keys": "配置键",
-    "No description": "暂无描述",
-    "config key": "配置键",
-    "Save config": "保存配置",
-    "New config": "新建配置",
-    "Cloud config saved.": "云配置已保存。",
-    "Invalid JSON value.": "JSON 值无效。",
-    "Loading invite audit...": "正在加载邀请审计...",
-    "Failed to load invite audit.": "邀请审计加载失败。",
-    "Code:": "邀请码：",
-    "IP:": "IP：",
-    "Device:": "设备：",
-    "Created at:": "创建时间：",
-    "Reason:": "原因：",
-    "Reject reward": "拒绝奖励",
-    "Reject reason": "拒绝原因",
-    "No invite redemptions matched the current filters.": "当前筛选条件下没有匹配的邀请兑换记录。",
-    "Revenue policy": "收益策略",
-    "Save policy": "保存策略",
-    "Loading policy.": "正在加载策略。",
-    "Contribution pool bps": "贡献池（基点）",
-    "Contribution window days": "贡献窗口（天）",
-    "Minimum settlement cents": "最低结算金额（分）",
-    "Event prices": "事件单价",
-    "Fixed shares": "固定分成",
-    "Contribution weights": "贡献权重",
-    "Display name": "显示名称",
-    "External ref id": "外部引用 ID",
-    Contact: "联系方式",
-    "Payout note": "结算备注",
-    "Save payee": "保存收益人",
-    "No payees.": "暂无收益人。",
-    "Settlement preview": "结算预览",
-    "Generate settlement": "生成结算",
-    Allocations: "分配条数",
-    "No allocations.": "暂无分配记录。",
-    "Revenue policy saved.": "收益策略已保存。",
-    "Payee saved.": "收益人已保存。",
-    "Settlement preview refreshed.": "结算预览已刷新。",
-    "Event prices JSON is invalid.": "事件单价 JSON 无效。",
-    "Fixed shares JSON is invalid.": "固定分成 JSON 无效。",
-    "Contribution weights JSON is invalid.": "贡献权重 JSON 无效。",
-    "Policy is not loaded.": "策略尚未加载。",
-
-    // Revenue Sharing enum values (rendered via t(value))
-    paused: "已暂停",
-    character: "角色",
-    world_owner: "世界所有者",
-    wiki_user: "Wiki 用户",
-    system: "系统",
-    provider: "供应方",
-    runtime_operator: "运行时运营方",
-    platform: "平台",
-    character_creator: "角色创建者",
-    character_editor: "角色编辑者",
-    character_reviewer: "角色审核者",
-    character_patroller: "角色巡检者",
-    logic_publisher: "逻辑发布者",
-    unassigned_hold: "未分配（暂扣）",
-    held: "暂扣",
-    payable: "应付",
-    settled: "已结算",
-    character_chat_message: "角色对话消息",
-    character_voice_turn: "角色语音回合",
-    character_video_turn: "角色视频回合",
-    character_content_use: "角色内容使用",
-    character_logic_run: "角色逻辑运行",
-    character_create: "创建角色",
-    character_content_edit_approved: "角色内容编辑通过",
-    character_logic_edit_approved: "角色逻辑编辑通过",
-    character_review_approved: "角色审核通过",
-    character_patrol: "角色巡检",
-    character_logic_publish: "角色逻辑发布",
-
-    // Nav hints / route meta (shared by sidebar + header)
-    "Accounts and expiry": "账号与到期",
-    "Pricing and access": "价格与访问",
-    "Trial and copy": "试用与文案",
-    "Rewards and risk": "奖励与风控",
-    "Client analytics": "客户端分析",
-    "Payees and ledgers": "收益人与账本",
-    "Inspect subscription history, invite rewards, and one cloud account's world linkage.":
-      "查看订阅记录、邀请奖励，以及该云账号绑定的世界。",
-    "Review cloud account status, subscription expiry, inviter attribution, and world linkage.":
-      "复核云账号状态、订阅到期时间、邀请归因和世界绑定。",
-    "Review reward redemptions, device and IP signals, and manual reversals.":
-      "复核奖励兑换、设备与 IP 信号，以及人工撤销操作。",
-
-    // Worlds / Jobs / Dashboard / World detail pages
-    "Docker compose snippet": "Docker Compose 片段",
-    "Escalated worlds": "已升级世界",
-    "Filter worlds by critical attention": "按严重关注筛选世界",
-    "Filter worlds by error instances": "按错误实例筛选世界",
-    "Filter worlds by ready status": "按就绪状态筛选世界",
-    "Filter worlds by running instances": "按运行中实例筛选世界",
-    "Filter worlds by stopped instances": "按停止实例筛选世界",
-    "Filter worlds by unassigned provider": "按未分配 Provider 筛选世界",
-    "Filter worlds by provider": "按 Provider 筛选世界",
-    "Heartbeat": "心跳",
-    "Launch config snapshot": "启动配置快照",
-    "Loading alert status...": "正在加载告警状态...",
-    "Loading bootstrap package...": "正在加载启动包...",
-    "Loading provider catalog...": "正在加载 Provider 目录...",
-    "Loading runtime status...": "正在加载运行时状态...",
-    "Loading world...": "正在加载世界...",
-    "No active drift or heartbeat issues right now.": "当前没有活跃的漂移或心跳问题。",
-    "No current alert. This world is below escalation thresholds.": "当前没有告警。该世界低于升级阈值。",
-    "No instance rows match the current filter set.": "当前筛选条件下没有匹配的实例行。",
-    "No jobs match the selected queue filter.": "当前队列筛选下没有匹配的任务。",
-    "No jobs recorded for this world yet.": "该世界尚未记录任何任务。",
-    "Open full queue": "打开完整队列",
-    "Ops notes": "运维备注",
-    "Outcome": "结果",
-    "Power": "电源",
-    "Provider key": "供应方键",
-    "Provider profile": "供应方配置",
-    "Provider region": "供应方区域",
-    "Provider zone": "供应方可用区",
-    "Provision strategy": "建世界策略",
-    "World API base URL": "世界 API 基础 URL",
-    "World admin URL": "世界管理后台 URL",
-
-    // Helper-function return strings
-    "Absent": "无实例",
-    "Related jobs": "相关任务",
-    "Provider error": "Provider 错误",
-    "Runtime drift": "运行时漂移",
-    "Sleep drift": "休眠漂移",
-    "Not leased": "未租约",
-    "Unleased": "未租约",
-    "World failed": "世界失败",
-    "Retry threshold": "重试阈值",
-    "Heartbeat duration": "心跳时长",
-    "Not escalated": "未升级",
-    "Phone is required.": "请填写手机号。",
-    "World name is required.": "请填写世界名称。",
-    "A ready world must include a world API base URL.": "处于就绪状态的世界必须设置世界 API 基础 URL。",
-  },
-  "ja-JP": {
-    "Yinjie Cloud Ops": "隠界クラウド運用",
-    "Cloud World Console": "クラウド世界コンソール",
-    "Manage phone-based world provisioning, wake old worlds, inspect lifecycle jobs, and track instance health from one place.":
-      "電話番号ベースの世界作成、古い世界の起動、ライフサイクルジョブの確認、インスタンス健全性の追跡を一か所で管理します。",
-    "Enter CLOUD_ADMIN_SECRET": "CLOUD_ADMIN_SECRET を入力",
-    "Admin access required": "管理シークレットが必要です",
-    "Enter CLOUD_ADMIN_SECRET to unlock the console. Cloud requests are paused until a secret is saved locally.":
-      "コンソールを解除するには CLOUD_ADMIN_SECRET を入力してください。ローカルに保存されるまでクラウドリクエストは停止されます。",
-    "CLOUD_ADMIN_SECRET is required.":
-      "CLOUD_ADMIN_SECRET を入力してください。",
-    "CLOUD_ADMIN_SECRET is invalid.":
-      "CLOUD_ADMIN_SECRET が無効です。",
-    "Cloud admin session is invalid or expired.":
-      "クラウド管理セッションが無効、または期限切れです。",
-    "Admin secret saved locally. Short-lived admin tokens will refresh automatically.":
-      "管理シークレットをローカルに保存しました。短命の管理トークンは自動更新されます。",
-    "Admin secret cleared.": "管理シークレットをクリアしました。",
-    "Admin secret saved locally. Console uses short-lived admin tokens.":
-      "管理シークレットはローカルに保存済みです。コンソールは短命の管理トークンを使用します。",
-    "Admin secret is missing.": "管理シークレットがありません。",
-    "Admin secret configured": "管理シークレット設定済み",
-    "Admin secret missing": "管理シークレット未設定",
-    "Cloud access": "クラウドアクセス",
-    "Cloud operations": "クラウド運用",
-    Navigation: "ナビゲーション",
-    More: "その他",
-    Setup: "設定",
-    "Request detail": "申請詳細",
-    "Availability and drift": "可用性とドリフト",
-    "Applications and handoffs": "申請と引き継ぎ",
-    "Instances and health": "インスタンスと健全性",
-    "Queue and leases": "キューとリース",
-    "Access audit": "アクセス監査",
-    "Durable tasks": "永続タスク",
-    "World request detail": "世界申請詳細",
-    "Review one phone-based world application and its delivery state.":
-      "電話番号ベースの世界申請と配信状態を確認します。",
-    "Review application status, projected world state, and manual delivery handoffs.":
-      "申請状態、予測世界状態、手動配信の引き継ぎを確認します。",
-    "World detail": "世界詳細",
-    "Inspect instance placement, runtime status, bootstrap material, and lifecycle jobs.":
-      "インスタンス配置、ランタイム状態、ブートストラップ資料、ライフサイクルジョブを確認します。",
-    "Track provider placement, power state, heartbeat freshness, and operator attention.":
-      "プロバイダー配置、電源状態、ハートビート鮮度、運用注意項目を追跡します。",
-    "Session audit": "セッション監査",
-    "Review admin access, risk groups, revocation paths, and source activity.":
-      "管理アクセス、リスクグループ、失効経路、ソース活動を確認します。",
-    "Durable recovery": "永続リカバリー",
-    "Replay exhausted compensation tasks and clear stale failures from one queue.":
-      "1 つのキューから再試行上限に達した補償タスクを再実行し、古い失敗をクリアします。",
-    "Monitor world availability, lifecycle jobs, request flow, and cloud runtime drift.":
-      "世界の可用性、ライフサイクルジョブ、申請フロー、クラウドランタイムのドリフトを監視します。",
-    "Request id": "リクエスト ID",
-
-    Dashboard: "ダッシュボード",
-    Requests: "申請",
-    Worlds: "世界",
-    Jobs: "ジョブ",
-    Sessions: "セッション",
-    "Waiting Sync": "待機同期",
-    "Fleet Dashboard": "フリートダッシュボード",
-    "World requests": "世界申請",
-    "Managed worlds": "管理中の世界",
-    "Lifecycle jobs": "ライフサイクルジョブ",
-    "Admin sessions": "管理セッション",
-    "Waiting session sync": "待機セッション同期",
-    "Request guidance": "申請ガイダンス",
-    "Lifecycle summary": "ライフサイクル概要",
-    "Bootstrap package": "ブートストラップパッケージ",
-    "Attention Queue": "対応キュー",
-    "Batch actions": "一括操作",
-
-    "Ready worlds": "準備完了の世界",
-    "Critical attention": "重大な対応",
-    "Recovery queued": "復旧キュー済み",
-    "Provider drift": "Provider の差分",
-    "Instance footprint": "インスタンス使用状況",
-    "Request alerts": "申請アラート",
-    "Active jobs": "アクティブジョブ",
-    "Failed jobs": "失敗ジョブ",
-    "Superseded jobs": "置換済みジョブ",
-    "Running jobs": "実行中ジョブ",
-    "Lease expired jobs": "リース期限切れジョブ",
-    "Delayed jobs": "遅延ジョブ",
-    "Pending approvals": "承認待ち",
-    "Provisioning handoffs": "プロビジョニング引き継ぎ",
-    "Rejected requests": "拒否済み申請",
-    "Disabled requests": "無効化申請",
-    "Recent requests": "最近の申請",
-    "Recent lifecycle jobs": "最近のライフサイクルジョブ",
-    "Recent failed jobs": "最近の失敗ジョブ",
-    "Recent superseded jobs": "最近の置換済みジョブ",
-    "No active jobs right now.": "現在アクティブなジョブはありません。",
-    "No failed jobs right now.": "現在失敗したジョブはありません。",
-    "No superseded jobs right now.": "現在置換済みジョブはありません。",
-    "No attention items right now.": "現在対応項目はありません。",
-    "No recent requests.": "最近の申請はありません。",
-
-    All: "すべて",
-    Unknown: "不明",
-    "Not available": "利用不可",
-    None: "なし",
-    "Not set": "未設定",
-    Save: "保存",
-    Edit: "編集",
-    Copy: "コピー",
-    Open: "開く",
-    Close: "閉じる",
-    Cancel: "キャンセル",
-    Confirm: "確認",
-    Clear: "クリア",
-    Reset: "リセット",
-    Retry: "再試行",
-    Refresh: "更新",
-    Apply: "適用",
-    Download: "ダウンロード",
-    Export: "エクスポート",
-    Search: "検索",
-    Submit: "送信",
-    Update: "更新",
-    Approve: "承認",
-    Reject: "拒否",
-    Disable: "無効化",
-    Enable: "有効化",
-    Resume: "再開",
-    Suspend: "停止",
-    Reconcile: "照合",
-    Rotate: "ローテート",
-    Revoke: "取り消し",
-    Hide: "非表示",
-    Show: "表示",
-    "Show more": "さらに表示",
-    "Show less": "折りたたむ",
-    "View all": "すべて表示",
-    "View jobs": "ジョブを見る",
-    "View world": "世界を見る",
-    "View worlds": "世界を見る",
-    "View requests": "申請を見る",
-    "View sessions": "セッションを見る",
-    "Open dashboard": "ダッシュボードを開く",
-    "Open requests": "申請を開く",
-    "Open worlds": "世界を開く",
-    "Open jobs": "ジョブを開く",
-    "Open sessions": "セッションを開く",
-    "Open waiting sync": "待機同期を開く",
-    "Open sessions permalink": "セッションの固定リンクを開く",
-    "Open requests permalink": "申請の固定リンクを開く",
-    "Open worlds permalink": "世界の固定リンクを開く",
-    "Open jobs permalink": "ジョブの固定リンクを開く",
-    "Open waiting-sync permalink": "待機同期の固定リンクを開く",
-
-    Status: "ステータス",
-    Phone: "電話番号",
-    "Phone number": "電話番号",
-    World: "世界",
-    Request: "申請",
-    Job: "ジョブ",
-    Session: "セッション",
-    "World id": "世界 ID",
-    "Job id": "ジョブ ID",
-    "Session id": "セッション ID",
-    Name: "名前",
-    Provider: "Provider",
-    Region: "リージョン",
-    Zone: "ゾーン",
-    Strategy: "戦略",
-    Health: "健全性",
-    "Desired state": "期待状態",
-    "Power state": "電源状態",
-    Attention: "対応",
-    Severity: "重要度",
-    Reason: "理由",
-    Message: "メッセージ",
-    Error: "エラー",
-    "Last error": "直近エラー",
-    "Failure code": "失敗コード",
-    "Failure message": "失敗メッセージ",
-    "API base URL": "API ベース URL",
-    "Admin URL": "管理 URL",
-    "Instance URL": "インスタンス URL",
-    "Callback token": "コールバックトークン",
-    "Runtime status": "ランタイム状態",
-    "Bootstrap config": "ブートストラップ設定",
-    "Docker compose": "Docker Compose",
-    "Runtime env overlay": "ランタイム環境オーバーレイ",
-    "Callback endpoints": "コールバックエンドポイント",
-    "Lifecycle actions": "ライフサイクル操作",
-    "Job queue": "ジョブキュー",
-    "Queue state": "キュー状態",
-    "Task type": "タスク種別",
-    "Task key": "タスクキー",
-    Target: "対象",
-    Context: "コンテキスト",
-    Attempt: "試行",
-    Available: "利用可能",
-    Updated: "更新",
-    Created: "作成",
-    Finished: "完了",
-    Started: "開始",
-    "Last accessed": "最終アクセス",
-    "Last interactive": "最終操作",
-    "Last booted": "最終起動",
-    "Last heartbeat": "最終ハートビート",
-    "Last suspended": "最終停止",
-    "Last operation": "最終操作",
-    "Observed at": "観測日時",
-    "Lease owner": "リース所有者",
-    "Lease expires": "リース期限",
-    "Latest update": "最新更新",
-    "Latest event": "最新イベント",
-    "Last generated": "最終生成",
-
-    Ready: "準備完了",
-    Pending: "保留中",
-    Provisioning: "作成中",
-    Active: "アクティブ",
-    Rejected: "拒否済み",
-    Disabled: "無効",
-    Failed: "失敗",
-    Succeeded: "成功",
-    Running: "実行中",
-    Queued: "キュー済み",
-    Sleeping: "スリープ中",
-    Stopping: "停止中",
-    Starting: "起動中",
-    Stopped: "停止済み",
-    Expired: "期限切れ",
-    Revoked: "取り消し済み",
-    Current: "現在",
-    Healthy: "正常",
-    Critical: "重大",
-    Warning: "警告",
-    Normal: "通常",
-    Watch: "監視",
-    "All statuses": "すべてのステータス",
-    "All reasons": "すべての理由",
-    "All sessions": "すべてのセッション",
-    "Current only": "現在のみ",
-    Ascending: "昇順",
-    Descending: "降順",
-    "Last used": "最終使用",
-    "Revoked at": "取り消し日時",
-    Logout: "ログアウト",
-    "Manual revocation": "手動取り消し",
-    "Refresh token reuse": "リフレッシュトークン再利用",
-    "Multiple active sessions": "複数のアクティブセッション",
-    "Repeated revocations": "繰り返し取り消し",
-
-    "Request search": "申請検索",
-    "World search": "世界検索",
-    "Waiting sync search": "待機同期検索",
-    "Projected world status": "予測世界ステータス",
-    "Projected desired state": "予測期待状態",
-    "Waiting sync status": "待機同期ステータス",
-    "Waiting sync task type": "待機同期タスク種別",
-    "Waiting sync page size": "待機同期ページサイズ",
-    "Session id, IP, client, revoker": "セッション ID、IP、クライアント、取り消し者",
-    "Search world, phone, job, lease...": "世界、電話番号、ジョブ、リースを検索...",
-    "task key, target, context, or error": "タスクキー、対象、コンテキスト、エラー",
-    "request id, phone, world name, or status detail":
-      "申請 ID、電話番号、世界名、ステータス詳細",
-    "world id, phone, name, provider, or endpoint":
-      "世界 ID、電話番号、名前、Provider、エンドポイント",
-    "world id, phone, email, name, provider, or endpoint":
-      "世界 ID、電話番号、メール、名前、Provider、エンドポイント",
-
-    "Copy callback token": "コールバックトークンをコピー",
-    "Copy runtime env overlay": "ランタイム環境オーバーレイをコピー",
-    "Copy docker compose snippet": "Docker Compose スニペットをコピー",
-    "Context task review": "コンテキストタスクレビュー",
-    "Recent task receipts": "最近のタスク受領",
-    "Unable to load waiting sync tasks": "待機同期タスクを読み込めません",
-    "Unable to load requests": "申請を読み込めません",
-    "Unable to load worlds": "世界を読み込めません",
-    "Unable to load jobs": "ジョブを読み込めません",
-    "Unable to load admin sessions": "管理セッションを読み込めません",
-
-    "Refresh world": "世界を更新",
-    "Refresh phone": "電話番号を更新",
-    "Invalidate phone": "電話番号を無効化",
-    "Replay failed": "失敗分を再実行",
-    "Clear failed": "失敗分をクリア",
-    "Replay task": "タスクを再実行",
-    "Clear task": "タスクをクリア",
-    "Batch replay failed": "失敗分を一括再実行",
-    "Batch clear failed": "失敗分を一括クリア",
-    "Filtered replay failed": "絞り込み失敗分を再実行",
-    "Filtered clear failed": "絞り込み失敗分をクリア",
-    "Switch status to All or Failed before running batch failed-task actions.":
-      "失敗タスクの一括操作を実行する前に、ステータスをすべてまたは失敗に切り替えてください。",
-    "All failed tasks across every page.": "全ページのすべての失敗タスク。",
-    "Focus snapshot appears when the current query exactly matches a visible context or target.":
-      "現在の検索が表示中のコンテキストまたは対象に完全一致すると、フォーカススナップショットが表示されます。",
-    "Add a context or target query to export a tighter investigation snapshot.":
-      "コンテキストまたは対象の検索条件を追加すると、より絞り込んだ調査スナップショットをエクスポートできます。",
-
-    "Rotate the callback token?": "コールバックトークンをローテートしますか？",
-    "Revoke admin session?": "管理セッションを取り消しますか？",
-    "Revoke selected admin sessions?": "選択した管理セッションを取り消しますか？",
-    "Revoke all matching admin sessions?": "一致するすべての管理セッションを取り消しますか？",
-    "Revoke matching risk groups?": "一致するリスクグループを取り消しますか？",
-    "Revoke source group?": "ソースグループを取り消しますか？",
-    "Close dialog": "ダイアログを閉じる",
-
-    "queue: all": "キュー: すべて",
-    "queue: running": "キュー: 実行中",
-    "queue: lease expired": "キュー: リース期限切れ",
-    "queue: delayed": "キュー: 遅延",
-    "queue: pending": "キュー: 保留中",
-    "sort by: updated": "並び替え: 更新",
-    "sort by: created": "並び替え: 作成",
-    "sort by: status": "並び替え: ステータス",
-    "direction: descending": "方向: 降順",
-    "direction: ascending": "方向: 昇順",
-
-    // SaaS / Revenue Sharing pages
-    Users: "ユーザー",
-    Plans: "プラン",
-    Configs: "設定",
-    "Invite Audit": "招待監査",
-    "Invite audit": "招待監査",
-    "Revenue Sharing": "収益シェア",
-    "SaaS operations": "SaaS 運用",
-    "Cloud monetization": "クラウド収益",
-    "User detail": "ユーザー詳細",
-    Account: "アカウント",
-    Subscription: "サブスクリプション",
-    Expires: "有効期限",
-    Inviter: "招待者",
-    Plan: "プラン",
-    Previous: "前へ",
-    Next: "次へ",
-    Inactive: "無効",
-    "Not loaded": "未読み込み",
-    Policy: "ポリシー",
-    Version: "バージョン",
-    Payable: "支払予定",
-    Held: "保留中",
-    Enabled: "有効化",
-    Currency: "通貨",
-    "External ref": "外部参照",
-    Payees: "受取人",
-    Payee: "受取人",
-    Participant: "参加者",
-    Amount: "金額",
-    Unassigned: "未割当",
-    Ledger: "元帳",
-    Events: "イベント",
-    Preview: "プレビュー",
-    "Usage type": "使用タイプ",
-    Character: "キャラクター",
-    Gross: "総額",
-    Occurred: "発生日時",
-    "Contribution type": "貢献タイプ",
-    Contributor: "貢献者",
-    State: "状態",
-    reversed: "取消済",
-    active: "アクティブ",
-    banned: "停止中",
-    archived: "アーカイブ済",
-    expired: "期限切れ",
-    none: "なし",
-    rewarded: "報酬付与済",
-    rejected: "却下済",
-    trial: "トライアル",
-    purchasable: "購入可能",
-    "All account states": "すべてのアカウント状態",
-    "All subscription states": "すべての購読状態",
-    "All redemption states": "すべての引換状態",
-    "Search phone": "電話番号で検索",
-    "Search phone or code": "電話番号または招待コードで検索",
-    "Loading SaaS users...": "SaaS ユーザーを読み込み中...",
-    "Failed to load users.": "ユーザーの読み込みに失敗しました。",
-    "No cloud users matched the current filters.":
-      "現在のフィルターに一致するクラウドユーザーはいません。",
-    "Loading cloud user...": "クラウドユーザーを読み込み中...",
-    "Failed to load cloud user.": "クラウドユーザーの読み込みに失敗しました。",
-    "SaaS user": "SaaS ユーザー",
-    "Account:": "アカウント:",
-    "Subscription:": "サブスクリプション:",
-    "Current plan:": "現在のプラン:",
-    "Expires at:": "有効期限:",
-    "Invite code:": "招待コード:",
-    "World status:": "ワールド状態:",
-    "World:": "ワールド:",
-    "Open world": "ワールドを開く",
-    "Back to users": "ユーザー一覧へ戻る",
-    "Manual grant": "手動付与",
-    "Grant days": "日数を付与",
-    "Account state": "アカウント状態",
-    "Ban reason": "停止理由",
-    Ban: "停止",
-    Unban: "停止解除",
-    "Subscription history": "サブスクリプション履歴",
-    "No subscription records found.": "サブスクリプション記録はありません。",
-    "Invite history": "招待履歴",
-    "No invite rewards recorded.": "招待報酬の記録はありません。",
-    "Loading subscription plans...": "サブスクリプションプランを読み込み中...",
-    "Failed to load subscription plans.": "サブスクリプションプランの読み込みに失敗しました。",
-    "Save plan": "プランを保存",
-    "New plan": "新規プラン",
-    "Subscription plan saved.": "サブスクリプションプランを保存しました。",
-    code: "コード",
-    name: "名称",
-    durationDays: "期間（日）",
-    priceCents: "価格（セント）",
-    currency: "通貨",
-    sortOrder: "並び順",
-    description: "説明",
-    "Loading cloud configs...": "クラウド設定を読み込み中...",
-    "Failed to load cloud configs.": "クラウド設定の読み込みに失敗しました。",
-    "Config keys": "設定キー",
-    "No description": "説明なし",
-    "config key": "設定キー",
-    "Save config": "設定を保存",
-    "New config": "新規設定",
-    "Cloud config saved.": "クラウド設定を保存しました。",
-    "Invalid JSON value.": "JSON 値が無効です。",
-    "Loading invite audit...": "招待監査を読み込み中...",
-    "Failed to load invite audit.": "招待監査の読み込みに失敗しました。",
-    "Code:": "コード:",
-    "IP:": "IP:",
-    "Device:": "デバイス:",
-    "Created at:": "作成日時:",
-    "Reason:": "理由:",
-    "Reject reward": "報酬を拒否",
-    "Reject reason": "拒否理由",
-    "No invite redemptions matched the current filters.":
-      "現在のフィルターに一致する招待引換はありません。",
-    "Revenue policy": "収益ポリシー",
-    "Save policy": "ポリシーを保存",
-    "Loading policy.": "ポリシーを読み込み中。",
-    "Contribution pool bps": "貢献プール（bps）",
-    "Contribution window days": "貢献ウィンドウ（日）",
-    "Minimum settlement cents": "最低決済額（セント）",
-    "Event prices": "イベント価格",
-    "Fixed shares": "固定シェア",
-    "Contribution weights": "貢献重み",
-    "Display name": "表示名",
-    "External ref id": "外部参照 ID",
-    Contact: "連絡先",
-    "Payout note": "支払メモ",
-    "Save payee": "受取人を保存",
-    "No payees.": "受取人はいません。",
-    "Settlement preview": "決済プレビュー",
-    "Generate settlement": "決済を生成",
-    Allocations: "割当件数",
-    "No allocations.": "割当はありません。",
-    "Revenue policy saved.": "収益ポリシーを保存しました。",
-    "Payee saved.": "受取人を保存しました。",
-    "Settlement preview refreshed.": "決済プレビューを更新しました。",
-    "Event prices JSON is invalid.": "イベント価格の JSON が無効です。",
-    "Fixed shares JSON is invalid.": "固定シェアの JSON が無効です。",
-    "Contribution weights JSON is invalid.": "貢献重みの JSON が無効です。",
-    "Policy is not loaded.": "ポリシーが読み込まれていません。",
-
-    // Revenue Sharing enum values
-    paused: "一時停止",
-    character: "キャラクター",
-    world_owner: "世界オーナー",
-    wiki_user: "Wiki ユーザー",
-    system: "システム",
-    provider: "プロバイダー",
-    runtime_operator: "運用担当",
-    platform: "プラットフォーム",
-    character_creator: "キャラクター作成者",
-    character_editor: "キャラクター編集者",
-    character_reviewer: "キャラクターレビュアー",
-    character_patroller: "キャラクター巡回者",
-    logic_publisher: "ロジック公開者",
-    unassigned_hold: "未割当（保留）",
-    held: "保留",
-    payable: "支払予定",
-    settled: "決済済み",
-    character_chat_message: "キャラクターチャット",
-    character_voice_turn: "キャラクター音声ターン",
-    character_video_turn: "キャラクター映像ターン",
-    character_content_use: "キャラクターコンテンツ使用",
-    character_logic_run: "キャラクターロジック実行",
-    character_create: "キャラクター作成",
-    character_content_edit_approved: "キャラクターコンテンツ編集承認",
-    character_logic_edit_approved: "キャラクターロジック編集承認",
-    character_review_approved: "キャラクターレビュー承認",
-    character_patrol: "キャラクター巡回",
-    character_logic_publish: "キャラクターロジック公開",
-
-    // Nav hints / route meta
-    "Accounts and expiry": "アカウントと有効期限",
-    "Pricing and access": "料金とアクセス",
-    "Trial and copy": "トライアルと文言",
-    "Rewards and risk": "報酬とリスク",
-    "Client analytics": "クライアント解析",
-    "Payees and ledgers": "受取人と元帳",
-    "Inspect subscription history, invite rewards, and one cloud account's world linkage.":
-      "サブスクリプション履歴、招待報酬、そのクラウドアカウントに紐づくワールドを確認します。",
-    "Review cloud account status, subscription expiry, inviter attribution, and world linkage.":
-      "クラウドアカウント状態、サブスクリプション期限、招待元帰属、ワールド紐付けを確認します。",
-    "Configure trial, monthly, quarterly, yearly, and invite reward plans shown to the app.":
-      "App に表示されるトライアル、月額、四半期、年額、招待報酬プランを設定します。",
-    "Manage SaaS trial, invite, feature, copy, and public app URL configuration.":
-      "SaaS のトライアル、招待、機能、文言、公開 App URL の設定を管理します。",
-    "Review reward redemptions, device and IP signals, and manual reversals.":
-      "報酬の引換、デバイスと IP のシグナル、手動撤回を確認します。",
-    "Configure role usage pricing, payees, contribution weights, and settlement ledgers.":
-      "キャラクター利用の価格、受取人、貢献重み、決済元帳を設定します。",
-
-    // Worlds / Jobs / Dashboard / World detail pages
-    "Docker compose snippet": "Docker Compose スニペット",
-    "Escalated worlds": "エスカレーション済みワールド",
-    "Filter worlds by critical attention": "重大対応でワールドを絞り込む",
-    "Filter worlds by error instances": "エラーインスタンスでワールドを絞り込む",
-    "Filter worlds by ready status": "準備完了状態でワールドを絞り込む",
-    "Filter worlds by running instances": "実行中インスタンスでワールドを絞り込む",
-    "Filter worlds by stopped instances": "停止インスタンスでワールドを絞り込む",
-    "Filter worlds by unassigned provider": "未割当 Provider でワールドを絞り込む",
-    "Filter worlds by provider": "Provider でワールドを絞り込む",
-    "Heartbeat": "ハートビート",
-    "Launch config snapshot": "起動設定スナップショット",
-    "Loading alert status...": "アラート状態を読み込み中...",
-    "Loading bootstrap package...": "ブートストラップパッケージを読み込み中...",
-    "Loading provider catalog...": "Provider カタログを読み込み中...",
-    "Loading runtime status...": "ランタイム状態を読み込み中...",
-    "Loading world...": "ワールドを読み込み中...",
-    "No active drift or heartbeat issues right now.": "現在、アクティブなドリフトやハートビートの問題はありません。",
-    "No current alert. This world is below escalation thresholds.": "現在アラートはありません。このワールドはエスカレーション閾値を下回っています。",
-    "No instance rows match the current filter set.": "現在のフィルターに一致するインスタンス行はありません。",
-    "No jobs match the selected queue filter.": "選択したキューフィルターに一致するジョブはありません。",
-    "No jobs recorded for this world yet.": "このワールドにはまだジョブの記録がありません。",
-    "Open full queue": "完全なキューを開く",
-    "Ops notes": "運用メモ",
-    "Outcome": "結果",
-    "Power": "電源",
-    "Provider key": "Provider キー",
-    "Provider profile": "Provider プロファイル",
-    "Provider region": "Provider 地域",
-    "Provider zone": "Provider ゾーン",
-    "Provision strategy": "プロビジョニング戦略",
-    "World API base URL": "ワールド API ベース URL",
-    "World admin URL": "ワールド管理 URL",
-
-    // Helper-function return strings
-    "Absent": "なし",
-    "Related jobs": "関連ジョブ",
-    "Provider error": "Provider エラー",
-    "Runtime drift": "ランタイムドリフト",
-    "Sleep drift": "スリープドリフト",
-    "Not leased": "未リース",
-    "Unleased": "未リース",
-    "World failed": "ワールド失敗",
-    "Retry threshold": "再試行閾値",
-    "Heartbeat duration": "ハートビート時間",
-    "Not escalated": "未エスカレーション",
-    "Phone is required.": "電話番号を入力してください。",
-    "World name is required.": "ワールド名を入力してください。",
-    "A ready world must include a world API base URL.": "準備完了状態のワールドにはワールド API ベース URL が必要です。",
-
-    "Access": "アクセス",
-    "Actions": "操作",
-    "Admin": "管理者",
-    "Alert status": "アラート状態",
-    "Attention worlds": "要対応ワールド",
-    "Bootstrapped": "ブート済み",
-    "Clear matching failed tasks": "一致する失敗タスクをクリア",
-    "Clear world scope": "ワールドスコープをクリア",
-    "Code": "コード",
-    "Config key": "設定キー",
-    "Container": "コンテナ",
-    "Copy compose": "compose をコピー",
-    "Copy endpoints": "エンドポイントをコピー",
-    "Copy env": "環境変数をコピー",
-    "Copy jobs permalink": "ジョブの固定リンクをコピー",
-    "Copy token": "トークンをコピー",
-    "Copy worlds permalink": "ワールドの固定リンクをコピー",
-    "Critical alerts": "重大アラート",
-    "Default region": "デフォルトリージョン",
-    "Default zone": "デフォルトゾーン",
-    "Deployment": "デプロイ",
-    "Deployment mode": "デプロイモード",
-    "Deployment state": "デプロイ状態",
-    "Description": "説明",
-    "Disk": "ディスク",
-    "Duration (days)": "期間（日数）",
-    "Escalated": "エスカレート済み",
-    "Executor mode": "エグゼキューターモード",
-    "Export context groups CSV": "コンテキストグループ CSV をエクスポート",
-    "Export context groups snapshot": "コンテキストグループスナップショットをエクスポート",
-    "Export filtered CSV": "フィルタ済み CSV をエクスポート",
-    "Export filtered snapshot": "フィルタ済みスナップショットをエクスポート",
-    "Export risk groups CSV": "リスクグループ CSV をエクスポート",
-    "Export risk sessions CSV": "リスクセッション CSV をエクスポート",
-    "Export risk snapshot": "リスクスナップショットをエクスポート",
-    "Export snapshot": "スナップショットをエクスポート",
-    "Flavor": "フレーバー",
-    "Health message": "ヘルスメッセージ",
-    "Heartbeat stale": "ハートビート古い",
-    "Image": "イメージ",
-    "Inspect jobs": "ジョブを確認",
-    "Instance": "インスタンス",
-    "Instance fleet": "インスタンスフリート",
-    "Instance pool": "インスタンスプール",
-    "Internal note shown next to the key.": "キーの横に表示される社内メモ。",
-    "Latest failed lifecycle work that may need manual recovery.": "手動リカバリが必要となり得る最新の失敗ライフサイクルタスクです。",
-    "Latest lifecycle jobs that were replaced by newer work.": "新しいタスクに置き換えられた最新のライフサイクルジョブです。",
-    "Lease": "リース",
-    "Loading admin sessions...": "管理セッションを読み込み中…",
-    "Loading drift summary...": "ドリフトサマリを読み込み中…",
-    "Loading fleet dashboard...": "フリートダッシュボードを読み込み中…",
-    "Loading instances...": "インスタンスを読み込み中…",
-    "Loading jobs...": "ジョブを読み込み中…",
-    "Loading lifecycle jobs...": "ライフサイクルジョブを読み込み中…",
-    "Loading source groups...": "ソースグループを読み込み中…",
-    "Loading superseded lifecycle jobs...": "置き換え済みライフサイクルジョブを読み込み中…",
-    "Loading worlds...": "ワールドを読み込み中…",
-    "Managed lifecycle": "マネージドライフサイクル",
-    "Managed provisioning": "マネージドプロビジョニング",
-    "Most urgent worlds needing manual inspection or follow-up.": "手動確認またはフォローアップが必要な最も緊急のワールドです。",
-    "Needs action": "対応が必要",
-    "Next page": "次のページ",
-    "No": "いいえ",
-    "No active attention items. The fleet currently looks healthy.": "アクティブな要対応項目はありません。フリートは現在正常です。",
-    "No admin sessions match this filter.": "このフィルタに一致する管理セッションはありません。",
-    "No instance": "インスタンスなし",
-    "No instance record exists yet. Provisioning will create one automatically.": "インスタンスレコードはまだありません。プロビジョニング時に自動作成されます。",
-    "No jobs match this filter.": "このフィルタに一致するジョブはありません。",
-    "No pending or running jobs in the operator queue.": "オペレーターキューに保留中または実行中のジョブはありません。",
-    "No provider allocation data yet.": "Provider 割当データはまだありません。",
-    "No recent failed jobs.": "最近の失敗ジョブはありません。",
-    "No recent superseded jobs.": "最近の置き換え済みジョブはありません。",
-    "No source groups match this filter.": "このフィルタに一致するソースグループはありません。",
-    "No worlds match this filter.": "このフィルタに一致するワールドはありません。",
-    "Open world fleet": "ワールドフリートを開く",
-    "Operator Queue": "オペレーターキュー",
-    "Ops note": "運用メモ",
-    "Pending and running lifecycle work across the instance fleet.": "インスタンスフリート全体で保留中および実行中のライフサイクル作業です。",
-    "Plan code": "プランコード",
-    "Plan name": "プラン名",
-    "Previous page": "前のページ",
-    "Price (cents)": "価格（セント）",
-    "Private IP": "プライベート IP",
-    "Project": "プロジェクト",
-    "Provider instance": "Provider インスタンス",
-    "Provider message": "Provider メッセージ",
-    "Provider snapshot": "Provider スナップショット",
-    "Provider spread": "Provider 分布",
-    "Provider volume": "Provider ボリューム",
-    "Public IP": "パブリック IP",
-    "Publicly purchasable": "公開購入可",
-    "Queuing replay...": "再実行をキューに追加中…",
-    "Raw status": "未加工ステータス",
-    "Recent Failures": "最近の失敗",
-    "Refresh status": "状態を更新",
-    "Refreshing...": "更新中…",
-    "Remote host": "リモートホスト",
-    "Remote path": "リモートパス",
-    "Replay matching failed tasks": "一致する失敗タスクを再実行",
-    "Result": "結果",
-    "Revoke all matching": "一致するものをすべて取り消し",
-    "Revoke group": "グループを取り消し",
-    "Revoke matching risk groups": "一致するリスクグループを取り消し",
-    "Revoke now": "今すぐ取り消し",
-    "Revoke selected": "選択を取り消し",
-    "Running now": "現在実行中",
-    "Runtime is not checking in within the configured stale window.": "ランタイムが設定された期限内に応答していません。",
-    "Runtime observation": "ランタイム観測",
-    "Save world": "ワールドを保存",
-    "Saving...": "保存中…",
-    "Search worlds": "ワールドを検索",
-    "Select active sessions on this page to revoke them in one action.": "このページのアクティブセッションを選択すると、一括で取り消せます。",
-    "Select provider": "Provider を選択",
-    "Shown to users in the app subscription page.": "アプリのサブスクリプションページでユーザーに表示されます。",
-    "Snapshots": "スナップショット",
-    "Sort order": "並び順",
-    "Summary counts reflect all jobs matching the current filters, not just this page.": "サマリの件数は現在のフィルタに一致する全ジョブを反映しており、このページのみではありません。",
-    "Superseded Queue": "置き換え済みキュー",
-    "Top attention items": "最優先要対応項目",
-    "Total results": "結果合計",
-    "Trial": "トライアル",
-    "Value (JSON)": "値（JSON）",
-    "Viewing sessions": "セッション表示中",
-    "Visible failed": "表示中の失敗",
-    "Visible pending": "表示中の保留",
-    "Visible rows": "表示中の行",
-    "Visible running": "表示中の実行",
-    "World drift summary": "ワールドドリフトサマリ",
-    "World failed jobs": "ワールド失敗ジョブ",
-    "World name": "ワールド名",
-    "World operator jobs": "ワールドオペレータージョブ",
-    "World scope": "ワールドスコープ",
-    "World superseded jobs": "ワールド置き換え済みジョブ",
-    "Worlds that currently need operator attention.": "現在オペレーターの対応が必要なワールド。",
-    "Yes": "はい",
-    "e.g. New user trial": "例：新規ユーザートライアル",
-    "e.g. trial": "例: trial",
-    "{count} active session(s) selected on this page.": "このページで {count} 件のアクティブセッションを選択中。",
-    "{count} jobs": "{count} 件のジョブ",
-  },
-  "ko-KR": {
-    "Yinjie Cloud Ops": "인계 클라우드 운영",
-    "Cloud World Console": "클라우드 월드 콘솔",
-    "Manage phone-based world provisioning, wake old worlds, inspect lifecycle jobs, and track instance health from one place.":
-      "전화번호 기반 월드 생성, 기존 월드 깨우기, 라이프사이클 작업 점검, 인스턴스 상태 추적을 한곳에서 관리합니다.",
-    "Enter CLOUD_ADMIN_SECRET": "CLOUD_ADMIN_SECRET 입력",
-    "Admin access required": "관리 시크릿 필요",
-    "Enter CLOUD_ADMIN_SECRET to unlock the console. Cloud requests are paused until a secret is saved locally.":
-      "콘솔을 열려면 CLOUD_ADMIN_SECRET을 입력하세요. 로컬에 저장될 때까지 클라우드 요청은 일시 중지됩니다.",
-    "CLOUD_ADMIN_SECRET is required.":
-      "CLOUD_ADMIN_SECRET을 먼저 입력하세요.",
-    "CLOUD_ADMIN_SECRET is invalid.":
-      "CLOUD_ADMIN_SECRET이 올바르지 않습니다.",
-    "Cloud admin session is invalid or expired.":
-      "클라우드 관리 세션이 유효하지 않거나 만료되었습니다.",
-    "Admin secret saved locally. Short-lived admin tokens will refresh automatically.":
-      "관리 시크릿이 로컬에 저장되었습니다. 단기 관리 토큰은 자동으로 갱신됩니다.",
-    "Admin secret cleared.": "관리 시크릿이 삭제되었습니다.",
-    "Admin secret saved locally. Console uses short-lived admin tokens.":
-      "관리 시크릿이 로컬에 저장되어 있습니다. 콘솔은 단기 관리 토큰을 사용합니다.",
-    "Admin secret is missing.": "관리 시크릿이 없습니다.",
-    "Admin secret configured": "관리 시크릿 설정됨",
-    "Admin secret missing": "관리 시크릿 없음",
-    "Cloud access": "클라우드 접근",
-    "Cloud operations": "클라우드 운영",
-    Navigation: "탐색",
-    More: "더보기",
-    Setup: "설정",
-    "Request detail": "요청 상세",
-    "Availability and drift": "가용성과 드리프트",
-    "Applications and handoffs": "요청과 인계",
-    "Instances and health": "인스턴스와 상태",
-    "Queue and leases": "큐와 리스",
-    "Access audit": "접근 감사",
-    "Durable tasks": "지속 작업",
-    "World request detail": "월드 요청 상세",
-    "Review one phone-based world application and its delivery state.":
-      "전화번호 기반 월드 요청 하나와 전달 상태를 검토합니다.",
-    "Review application status, projected world state, and manual delivery handoffs.":
-      "요청 상태, 예상 월드 상태, 수동 전달 인계를 검토합니다.",
-    "World detail": "월드 상세",
-    "Inspect instance placement, runtime status, bootstrap material, and lifecycle jobs.":
-      "인스턴스 배치, 런타임 상태, 부트스트랩 자료, 라이프사이클 작업을 점검합니다.",
-    "Track provider placement, power state, heartbeat freshness, and operator attention.":
-      "공급자 배치, 전원 상태, 하트비트 최신성, 운영자 관심 항목을 추적합니다.",
-    "Session audit": "세션 감사",
-    "Review admin access, risk groups, revocation paths, and source activity.":
-      "관리 접근, 위험 그룹, 해지 경로, 소스 활동을 검토합니다.",
-    "Durable recovery": "지속 복구",
-    "Replay exhausted compensation tasks and clear stale failures from one queue.":
-      "하나의 큐에서 재시도 한도에 도달한 보상 작업을 재실행하고 오래된 실패를 정리합니다.",
-    "Monitor world availability, lifecycle jobs, request flow, and cloud runtime drift.":
-      "월드 가용성, 라이프사이클 작업, 요청 흐름, 클라우드 런타임 드리프트를 모니터링합니다.",
-    "Request id": "요청 ID",
-
-    Dashboard: "대시보드",
-    Requests: "요청",
-    Worlds: "월드",
-    Jobs: "작업",
-    Sessions: "세션",
-    "Waiting Sync": "대기 동기화",
-    "Fleet Dashboard": "플릿 대시보드",
-    "World requests": "월드 요청",
-    "Managed worlds": "관리 중인 월드",
-    "Lifecycle jobs": "라이프사이클 작업",
-    "Admin sessions": "관리 세션",
-    "Waiting session sync": "대기 세션 동기화",
-    "Request guidance": "요청 가이드",
-    "Lifecycle summary": "라이프사이클 요약",
-    "Bootstrap package": "부트스트랩 패키지",
-    "Attention Queue": "관심 큐",
-    "Batch actions": "일괄 작업",
-
-    "Ready worlds": "준비된 월드",
-    "Critical attention": "긴급 관심",
-    "Recovery queued": "복구 대기열",
-    "Provider drift": "Provider 드리프트",
-    "Instance footprint": "인스턴스 사용 현황",
-    "Request alerts": "요청 알림",
-    "Active jobs": "활성 작업",
-    "Failed jobs": "실패한 작업",
-    "Superseded jobs": "대체된 작업",
-    "Running jobs": "실행 중 작업",
-    "Lease expired jobs": "리스 만료 작업",
-    "Delayed jobs": "지연 작업",
-    "Pending approvals": "승인 대기",
-    "Provisioning handoffs": "프로비저닝 인계",
-    "Rejected requests": "거절된 요청",
-    "Disabled requests": "비활성 요청",
-    "Recent requests": "최근 요청",
-    "Recent lifecycle jobs": "최근 라이프사이클 작업",
-    "Recent failed jobs": "최근 실패 작업",
-    "Recent superseded jobs": "최근 대체 작업",
-    "No active jobs right now.": "현재 활성 작업이 없습니다.",
-    "No failed jobs right now.": "현재 실패한 작업이 없습니다.",
-    "No superseded jobs right now.": "현재 대체된 작업이 없습니다.",
-    "No attention items right now.": "현재 관심 항목이 없습니다.",
-    "No recent requests.": "최근 요청이 없습니다.",
-
-    All: "전체",
-    Unknown: "알 수 없음",
-    "Not available": "사용 불가",
-    None: "없음",
-    "Not set": "미설정",
-    Save: "저장",
-    Edit: "편집",
-    Copy: "복사",
-    Open: "열기",
-    Close: "닫기",
-    Cancel: "취소",
-    Confirm: "확인",
-    Clear: "지우기",
-    Reset: "초기화",
-    Retry: "다시 시도",
-    Refresh: "새로고침",
-    Apply: "적용",
-    Download: "다운로드",
-    Export: "내보내기",
-    Search: "검색",
-    Submit: "제출",
-    Update: "업데이트",
-    Approve: "승인",
-    Reject: "거절",
-    Disable: "비활성화",
-    Enable: "활성화",
-    Resume: "재개",
-    Suspend: "중지",
-    Reconcile: "조정",
-    Rotate: "교체",
-    Revoke: "해지",
-    Hide: "숨기기",
-    Show: "보기",
-    "Show more": "더 보기",
-    "Show less": "접기",
-    "View all": "전체 보기",
-    "View jobs": "작업 보기",
-    "View world": "월드 보기",
-    "View worlds": "월드 보기",
-    "View requests": "요청 보기",
-    "View sessions": "세션 보기",
-    "Open dashboard": "대시보드 열기",
-    "Open requests": "요청 열기",
-    "Open worlds": "월드 열기",
-    "Open jobs": "작업 열기",
-    "Open sessions": "세션 열기",
-    "Open waiting sync": "대기 동기화 열기",
-    "Open sessions permalink": "세션 고정 링크 열기",
-    "Open requests permalink": "요청 고정 링크 열기",
-    "Open worlds permalink": "월드 고정 링크 열기",
-    "Open jobs permalink": "작업 고정 링크 열기",
-    "Open waiting-sync permalink": "대기 동기화 고정 링크 열기",
-
-    Status: "상태",
-    Phone: "전화번호",
-    "Phone number": "전화번호",
-    World: "월드",
-    Request: "요청",
-    Job: "작업",
-    Session: "세션",
-    "World id": "월드 ID",
-    "Job id": "작업 ID",
-    "Session id": "세션 ID",
-    Name: "이름",
-    Provider: "Provider",
-    Region: "리전",
-    Zone: "존",
-    Strategy: "전략",
-    Health: "상태",
-    "Desired state": "희망 상태",
-    "Power state": "전원 상태",
-    Attention: "관심",
-    Severity: "심각도",
-    Reason: "이유",
-    Message: "메시지",
-    Error: "오류",
-    "Last error": "최근 오류",
-    "Failure code": "실패 코드",
-    "Failure message": "실패 메시지",
-    "API base URL": "API 기본 URL",
-    "Admin URL": "관리 URL",
-    "Instance URL": "인스턴스 URL",
-    "Callback token": "콜백 토큰",
-    "Runtime status": "런타임 상태",
-    "Bootstrap config": "부트스트랩 설정",
-    "Docker compose": "Docker Compose",
-    "Runtime env overlay": "런타임 환경 오버레이",
-    "Callback endpoints": "콜백 엔드포인트",
-    "Lifecycle actions": "라이프사이클 작업",
-    "Job queue": "작업 큐",
-    "Queue state": "큐 상태",
-    "Task type": "작업 유형",
-    "Task key": "작업 키",
-    Target: "대상",
-    Context: "컨텍스트",
-    Attempt: "시도",
-    Available: "사용 가능",
-    Updated: "업데이트",
-    Created: "생성",
-    Finished: "완료",
-    Started: "시작",
-    "Last accessed": "최근 접근",
-    "Last interactive": "최근 상호작용",
-    "Last booted": "최근 부팅",
-    "Last heartbeat": "최근 하트비트",
-    "Last suspended": "최근 중지",
-    "Last operation": "최근 작업",
-    "Observed at": "관측 시간",
-    "Lease owner": "리스 소유자",
-    "Lease expires": "리스 만료",
-    "Latest update": "최근 업데이트",
-    "Latest event": "최근 이벤트",
-    "Last generated": "최근 생성",
-
-    Ready: "준비됨",
-    Pending: "대기 중",
-    Provisioning: "생성 중",
-    Active: "활성",
-    Rejected: "거절됨",
-    Disabled: "비활성",
-    Failed: "실패",
-    Succeeded: "성공",
-    Running: "실행 중",
-    Queued: "대기열",
-    Sleeping: "절전",
-    Stopping: "중지 중",
-    Starting: "시작 중",
-    Stopped: "중지됨",
-    Expired: "만료됨",
-    Revoked: "해지됨",
-    Current: "현재",
-    Healthy: "정상",
-    Critical: "긴급",
-    Warning: "경고",
-    Normal: "정상",
-    Watch: "관찰",
-    "All statuses": "전체 상태",
-    "All reasons": "전체 이유",
-    "All sessions": "전체 세션",
-    "Current only": "현재만",
-    Ascending: "오름차순",
-    Descending: "내림차순",
-    "Last used": "최근 사용",
-    "Revoked at": "해지 시간",
-    Logout: "로그아웃",
-    "Manual revocation": "수동 해지",
-    "Refresh token reuse": "리프레시 토큰 재사용",
-    "Multiple active sessions": "여러 활성 세션",
-    "Repeated revocations": "반복 해지",
-
-    "Request search": "요청 검색",
-    "World search": "월드 검색",
-    "Waiting sync search": "대기 동기화 검색",
-    "Projected world status": "예상 월드 상태",
-    "Projected desired state": "예상 희망 상태",
-    "Waiting sync status": "대기 동기화 상태",
-    "Waiting sync task type": "대기 동기화 작업 유형",
-    "Waiting sync page size": "대기 동기화 페이지 크기",
-    "Session id, IP, client, revoker": "세션 ID, IP, 클라이언트, 해지자",
-    "Search world, phone, job, lease...": "월드, 전화번호, 작업, 리스 검색...",
-    "task key, target, context, or error": "작업 키, 대상, 컨텍스트 또는 오류",
-    "request id, phone, world name, or status detail":
-      "요청 ID, 전화번호, 월드 이름 또는 상태 세부 정보",
-    "world id, phone, name, provider, or endpoint":
-      "월드 ID, 전화번호, 이름, Provider 또는 엔드포인트",
-    "world id, phone, email, name, provider, or endpoint":
-      "월드 ID, 전화번호, 이메일, 이름, Provider 또는 엔드포인트",
-
-    "Copy callback token": "콜백 토큰 복사",
-    "Copy runtime env overlay": "런타임 환경 오버레이 복사",
-    "Copy docker compose snippet": "Docker Compose 스니펫 복사",
-    "Context task review": "컨텍스트 작업 검토",
-    "Recent task receipts": "최근 작업 영수증",
-    "Unable to load waiting sync tasks": "대기 동기화 작업을 불러올 수 없습니다",
-    "Unable to load requests": "요청을 불러올 수 없습니다",
-    "Unable to load worlds": "월드를 불러올 수 없습니다",
-    "Unable to load jobs": "작업을 불러올 수 없습니다",
-    "Unable to load admin sessions": "관리 세션을 불러올 수 없습니다",
-
-    "Refresh world": "월드 새로고침",
-    "Refresh phone": "전화번호 새로고침",
-    "Invalidate phone": "전화번호 무효화",
-    "Replay failed": "실패 항목 재실행",
-    "Clear failed": "실패 항목 지우기",
-    "Replay task": "작업 재실행",
-    "Clear task": "작업 지우기",
-    "Batch replay failed": "실패 항목 일괄 재실행",
-    "Batch clear failed": "실패 항목 일괄 지우기",
-    "Filtered replay failed": "필터된 실패 항목 재실행",
-    "Filtered clear failed": "필터된 실패 항목 지우기",
-    "Switch status to All or Failed before running batch failed-task actions.":
-      "실패 작업 일괄 작업을 실행하기 전에 상태를 전체 또는 실패로 바꾸세요.",
-    "All failed tasks across every page.": "모든 페이지의 모든 실패 작업입니다.",
-    "Focus snapshot appears when the current query exactly matches a visible context or target.":
-      "현재 검색어가 보이는 컨텍스트 또는 대상과 정확히 일치하면 포커스 스냅샷이 표시됩니다.",
-    "Add a context or target query to export a tighter investigation snapshot.":
-      "컨텍스트 또는 대상 검색어를 추가하면 더 좁은 조사 스냅샷을 내보낼 수 있습니다.",
-
-    "Rotate the callback token?": "콜백 토큰을 교체할까요?",
-    "Revoke admin session?": "관리 세션을 해지할까요?",
-    "Revoke selected admin sessions?": "선택한 관리 세션을 해지할까요?",
-    "Revoke all matching admin sessions?": "일치하는 모든 관리 세션을 해지할까요?",
-    "Revoke matching risk groups?": "일치하는 위험 그룹을 해지할까요?",
-    "Revoke source group?": "소스 그룹을 해지할까요?",
-    "Close dialog": "대화상자 닫기",
-
-    "queue: all": "큐: 전체",
-    "queue: running": "큐: 실행 중",
-    "queue: lease expired": "큐: 리스 만료",
-    "queue: delayed": "큐: 지연",
-    "queue: pending": "큐: 대기 중",
-    "sort by: updated": "정렬: 업데이트",
-    "sort by: created": "정렬: 생성",
-    "sort by: status": "정렬: 상태",
-    "direction: descending": "방향: 내림차순",
-    "direction: ascending": "방향: 오름차순",
-
-    // SaaS / Revenue Sharing pages
-    Users: "사용자",
-    Plans: "플랜",
-    Configs: "설정",
-    "Invite Audit": "초대 감사",
-    "Invite audit": "초대 감사",
-    "Revenue Sharing": "수익 분배",
-    "SaaS operations": "SaaS 운영",
-    "Cloud monetization": "클라우드 수익화",
-    "User detail": "사용자 상세",
-    Account: "계정",
-    Subscription: "구독",
-    Expires: "만료일",
-    Inviter: "초대자",
-    Plan: "플랜",
-    Previous: "이전",
-    Next: "다음",
-    Inactive: "비활성",
-    "Not loaded": "로드되지 않음",
-    Policy: "정책",
-    Version: "버전",
-    Payable: "지급 예정",
-    Held: "보류",
-    Enabled: "활성화",
-    Currency: "통화",
-    "External ref": "외부 참조",
-    Payees: "수취인",
-    Payee: "수취인",
-    Participant: "참여자",
-    Amount: "금액",
-    Unassigned: "할당되지 않음",
-    Ledger: "원장",
-    Events: "이벤트",
-    Preview: "미리보기",
-    "Usage type": "사용 유형",
-    Character: "캐릭터",
-    Gross: "총액",
-    Occurred: "발생일시",
-    "Contribution type": "기여 유형",
-    Contributor: "기여자",
-    State: "상태",
-    reversed: "취소됨",
-    active: "활성",
-    banned: "차단됨",
-    archived: "보관됨",
-    expired: "만료됨",
-    none: "없음",
-    rewarded: "보상됨",
-    rejected: "거부됨",
-    trial: "체험",
-    purchasable: "구매 가능",
-    "All account states": "모든 계정 상태",
-    "All subscription states": "모든 구독 상태",
-    "All redemption states": "모든 사용 상태",
-    "Search phone": "전화번호 검색",
-    "Search phone or code": "전화번호 또는 초대 코드 검색",
-    "Loading SaaS users...": "SaaS 사용자를 로드하는 중...",
-    "Failed to load users.": "사용자 로드에 실패했습니다.",
-    "No cloud users matched the current filters.":
-      "현재 필터와 일치하는 클라우드 사용자가 없습니다.",
-    "Loading cloud user...": "클라우드 사용자를 로드하는 중...",
-    "Failed to load cloud user.": "클라우드 사용자 로드에 실패했습니다.",
-    "SaaS user": "SaaS 사용자",
-    "Account:": "계정:",
-    "Subscription:": "구독:",
-    "Current plan:": "현재 플랜:",
-    "Expires at:": "만료일:",
-    "Invite code:": "초대 코드:",
-    "World status:": "월드 상태:",
-    "World:": "월드:",
-    "Open world": "월드 열기",
-    "Back to users": "사용자 목록으로",
-    "Manual grant": "수동 부여",
-    "Grant days": "일수 부여",
-    "Account state": "계정 상태",
-    "Ban reason": "차단 사유",
-    Ban: "차단",
-    Unban: "차단 해제",
-    "Subscription history": "구독 이력",
-    "No subscription records found.": "구독 기록이 없습니다.",
-    "Invite history": "초대 이력",
-    "No invite rewards recorded.": "초대 보상 기록이 없습니다.",
-    "Loading subscription plans...": "구독 플랜을 로드하는 중...",
-    "Failed to load subscription plans.": "구독 플랜 로드에 실패했습니다.",
-    "Save plan": "플랜 저장",
-    "New plan": "새 플랜",
-    "Subscription plan saved.": "구독 플랜을 저장했습니다.",
-    code: "코드",
-    name: "이름",
-    durationDays: "기간(일)",
-    priceCents: "가격(센트)",
-    currency: "통화",
-    sortOrder: "정렬 순서",
-    description: "설명",
-    "Loading cloud configs...": "클라우드 설정을 로드하는 중...",
-    "Failed to load cloud configs.": "클라우드 설정 로드에 실패했습니다.",
-    "Config keys": "설정 키",
-    "No description": "설명 없음",
-    "config key": "설정 키",
-    "Save config": "설정 저장",
-    "New config": "새 설정",
-    "Cloud config saved.": "클라우드 설정을 저장했습니다.",
-    "Invalid JSON value.": "JSON 값이 잘못되었습니다.",
-    "Loading invite audit...": "초대 감사를 로드하는 중...",
-    "Failed to load invite audit.": "초대 감사 로드에 실패했습니다.",
-    "Code:": "코드:",
-    "IP:": "IP:",
-    "Device:": "기기:",
-    "Created at:": "생성일시:",
-    "Reason:": "사유:",
-    "Reject reward": "보상 거부",
-    "Reject reason": "거부 사유",
-    "No invite redemptions matched the current filters.":
-      "현재 필터와 일치하는 초대 사용 기록이 없습니다.",
-    "Revenue policy": "수익 정책",
-    "Save policy": "정책 저장",
-    "Loading policy.": "정책을 로드하는 중.",
-    "Contribution pool bps": "기여 풀(bps)",
-    "Contribution window days": "기여 기간(일)",
-    "Minimum settlement cents": "최소 정산 금액(센트)",
-    "Event prices": "이벤트 단가",
-    "Fixed shares": "고정 분배",
-    "Contribution weights": "기여 가중치",
-    "Display name": "표시 이름",
-    "External ref id": "외부 참조 ID",
-    Contact: "연락처",
-    "Payout note": "정산 메모",
-    "Save payee": "수취인 저장",
-    "No payees.": "수취인이 없습니다.",
-    "Settlement preview": "정산 미리보기",
-    "Generate settlement": "정산 생성",
-    Allocations: "할당 건수",
-    "No allocations.": "할당이 없습니다.",
-    "Revenue policy saved.": "수익 정책을 저장했습니다.",
-    "Payee saved.": "수취인을 저장했습니다.",
-    "Settlement preview refreshed.": "정산 미리보기를 새로고침했습니다.",
-    "Event prices JSON is invalid.": "이벤트 단가 JSON이 잘못되었습니다.",
-    "Fixed shares JSON is invalid.": "고정 분배 JSON이 잘못되었습니다.",
-    "Contribution weights JSON is invalid.": "기여 가중치 JSON이 잘못되었습니다.",
-    "Policy is not loaded.": "정책이 로드되지 않았습니다.",
-
-    // Revenue Sharing enum values
-    paused: "일시 중지",
-    character: "캐릭터",
-    world_owner: "월드 소유자",
-    wiki_user: "Wiki 사용자",
-    system: "시스템",
-    provider: "공급자",
-    runtime_operator: "런타임 운영자",
-    platform: "플랫폼",
-    character_creator: "캐릭터 작성자",
-    character_editor: "캐릭터 편집자",
-    character_reviewer: "캐릭터 리뷰어",
-    character_patroller: "캐릭터 순찰자",
-    logic_publisher: "로직 게시자",
-    unassigned_hold: "미할당(보류)",
-    held: "보류",
-    payable: "지급 예정",
-    settled: "결산 완료",
-    character_chat_message: "캐릭터 채팅 메시지",
-    character_voice_turn: "캐릭터 음성 턴",
-    character_video_turn: "캐릭터 영상 턴",
-    character_content_use: "캐릭터 콘텐츠 사용",
-    character_logic_run: "캐릭터 로직 실행",
-    character_create: "캐릭터 생성",
-    character_content_edit_approved: "캐릭터 콘텐츠 편집 승인",
-    character_logic_edit_approved: "캐릭터 로직 편집 승인",
-    character_review_approved: "캐릭터 검토 승인",
-    character_patrol: "캐릭터 순찰",
-    character_logic_publish: "캐릭터 로직 공개",
-
-    // Nav hints / route meta
-    "Accounts and expiry": "계정과 만료",
-    "Pricing and access": "가격과 접근",
-    "Trial and copy": "체험과 문구",
-    "Rewards and risk": "보상과 리스크",
-    "Client analytics": "클라이언트 분석",
-    "Payees and ledgers": "수취인과 원장",
-    "Inspect subscription history, invite rewards, and one cloud account's world linkage.":
-      "구독 이력, 초대 보상, 해당 클라우드 계정의 월드 연동을 확인합니다.",
-    "Review cloud account status, subscription expiry, inviter attribution, and world linkage.":
-      "클라우드 계정 상태, 구독 만료, 초대자 귀속, 월드 연동을 확인합니다.",
-    "Configure trial, monthly, quarterly, yearly, and invite reward plans shown to the app.":
-      "App에 표시되는 체험, 월간, 분기, 연간, 초대 보상 플랜을 설정합니다.",
-    "Manage SaaS trial, invite, feature, copy, and public app URL configuration.":
-      "SaaS 체험, 초대, 기능, 문구, 공개 App URL 설정을 관리합니다.",
-    "Review reward redemptions, device and IP signals, and manual reversals.":
-      "보상 사용, 기기 및 IP 신호, 수동 취소를 검토합니다.",
-    "Configure role usage pricing, payees, contribution weights, and settlement ledgers.":
-      "역할 사용 가격, 수취인, 기여 가중치, 정산 원장을 설정합니다.",
-
-    // Worlds / Jobs / Dashboard / World detail pages
-    "Docker compose snippet": "Docker Compose 스니펫",
-    "Escalated worlds": "에스컬레이션된 월드",
-    "Filter worlds by critical attention": "심각 주의로 월드 필터링",
-    "Filter worlds by error instances": "오류 인스턴스로 월드 필터링",
-    "Filter worlds by ready status": "준비 완료 상태로 월드 필터링",
-    "Filter worlds by running instances": "실행 중 인스턴스로 월드 필터링",
-    "Filter worlds by stopped instances": "정지 인스턴스로 월드 필터링",
-    "Filter worlds by unassigned provider": "할당되지 않은 Provider로 월드 필터링",
-    "Filter worlds by provider": "Provider로 월드 필터링",
-    "Heartbeat": "하트비트",
-    "Launch config snapshot": "시작 설정 스냅샷",
-    "Loading alert status...": "경보 상태를 로드하는 중...",
-    "Loading bootstrap package...": "부트스트랩 패키지를 로드하는 중...",
-    "Loading provider catalog...": "Provider 카탈로그를 로드하는 중...",
-    "Loading runtime status...": "런타임 상태를 로드하는 중...",
-    "Loading world...": "월드를 로드하는 중...",
-    "No active drift or heartbeat issues right now.": "현재 활성 드리프트나 하트비트 문제가 없습니다.",
-    "No current alert. This world is below escalation thresholds.": "현재 경보가 없습니다. 이 월드는 에스컬레이션 임계값 이하입니다.",
-    "No instance rows match the current filter set.": "현재 필터와 일치하는 인스턴스 행이 없습니다.",
-    "No jobs match the selected queue filter.": "선택한 큐 필터와 일치하는 작업이 없습니다.",
-    "No jobs recorded for this world yet.": "이 월드에 기록된 작업이 아직 없습니다.",
-    "Open full queue": "전체 큐 열기",
-    "Ops notes": "운영 노트",
-    "Outcome": "결과",
-    "Power": "전원",
-    "Provider key": "Provider 키",
-    "Provider profile": "Provider 프로필",
-    "Provider region": "Provider 지역",
-    "Provider zone": "Provider 영역",
-    "Provision strategy": "프로비저닝 전략",
-    "World API base URL": "월드 API 기본 URL",
-    "World admin URL": "월드 관리자 URL",
-
-    // Helper-function return strings
-    "Absent": "없음",
-    "Related jobs": "관련 작업",
-    "Provider error": "Provider 오류",
-    "Runtime drift": "런타임 드리프트",
-    "Sleep drift": "슬립 드리프트",
-    "Not leased": "리스 안 됨",
-    "Unleased": "리스 안 됨",
-    "World failed": "월드 실패",
-    "Retry threshold": "재시도 임계값",
-    "Heartbeat duration": "하트비트 기간",
-    "Not escalated": "에스컬레이션 안 됨",
-    "Phone is required.": "전화번호를 입력하세요.",
-    "World name is required.": "월드 이름을 입력하세요.",
-    "A ready world must include a world API base URL.": "준비 완료 상태의 월드에는 월드 API 기본 URL이 필요합니다.",
-
-    "Access": "접근",
-    "Actions": "작업",
-    "Admin": "관리자",
-    "Alert status": "알림 상태",
-    "Attention worlds": "주의 월드",
-    "Bootstrapped": "부트스트랩 완료",
-    "Clear matching failed tasks": "일치하는 실패 작업 정리",
-    "Clear world scope": "월드 범위 지우기",
-    "Code": "코드",
-    "Config key": "설정 키",
-    "Container": "컨테이너",
-    "Copy compose": "compose 복사",
-    "Copy endpoints": "엔드포인트 복사",
-    "Copy env": "환경변수 복사",
-    "Copy jobs permalink": "작업 고정 링크 복사",
-    "Copy token": "토큰 복사",
-    "Copy worlds permalink": "월드 고정 링크 복사",
-    "Critical alerts": "치명적 알림",
-    "Default region": "기본 리전",
-    "Default zone": "기본 영역",
-    "Deployment": "배포",
-    "Deployment mode": "배포 모드",
-    "Deployment state": "배포 상태",
-    "Description": "설명",
-    "Disk": "디스크",
-    "Duration (days)": "기간(일)",
-    "Escalated": "에스컬레이션됨",
-    "Executor mode": "실행기 모드",
-    "Export context groups CSV": "컨텍스트 그룹 CSV 내보내기",
-    "Export context groups snapshot": "컨텍스트 그룹 스냅샷 내보내기",
-    "Export filtered CSV": "필터링된 CSV 내보내기",
-    "Export filtered snapshot": "필터링된 스냅샷 내보내기",
-    "Export risk groups CSV": "위험 그룹 CSV 내보내기",
-    "Export risk sessions CSV": "위험 세션 CSV 내보내기",
-    "Export risk snapshot": "위험 스냅샷 내보내기",
-    "Export snapshot": "스냅샷 내보내기",
-    "Flavor": "플레이버",
-    "Health message": "상태 메시지",
-    "Heartbeat stale": "하트비트 만료",
-    "Image": "이미지",
-    "Inspect jobs": "작업 검사",
-    "Instance": "인스턴스",
-    "Instance fleet": "인스턴스 플릿",
-    "Instance pool": "인스턴스 풀",
-    "Internal note shown next to the key.": "키 옆에 표시되는 내부 메모입니다.",
-    "Latest failed lifecycle work that may need manual recovery.": "수동 복구가 필요할 수 있는 최근 실패 라이프사이클 작업입니다.",
-    "Latest lifecycle jobs that were replaced by newer work.": "새 작업으로 대체된 최근 라이프사이클 작업입니다.",
-    "Lease": "리스",
-    "Loading admin sessions...": "관리자 세션을 불러오는 중…",
-    "Loading drift summary...": "드리프트 요약을 불러오는 중…",
-    "Loading fleet dashboard...": "플릿 대시보드를 불러오는 중…",
-    "Loading instances...": "인스턴스를 불러오는 중…",
-    "Loading jobs...": "작업을 불러오는 중…",
-    "Loading lifecycle jobs...": "라이프사이클 작업을 불러오는 중…",
-    "Loading source groups...": "소스 그룹을 불러오는 중…",
-    "Loading superseded lifecycle jobs...": "대체된 라이프사이클 작업을 불러오는 중…",
-    "Loading worlds...": "월드를 불러오는 중…",
-    "Managed lifecycle": "관리형 라이프사이클",
-    "Managed provisioning": "관리형 프로비저닝",
-    "Most urgent worlds needing manual inspection or follow-up.": "수동 확인이나 후속 조치가 가장 시급한 월드입니다.",
-    "Needs action": "조치 필요",
-    "Next page": "다음 페이지",
-    "No": "아니오",
-    "No active attention items. The fleet currently looks healthy.": "활성 주의 항목이 없습니다. 플릿은 현재 정상 상태입니다.",
-    "No admin sessions match this filter.": "이 필터에 해당하는 관리자 세션이 없습니다.",
-    "No instance": "인스턴스 없음",
-    "No instance record exists yet. Provisioning will create one automatically.": "아직 인스턴스 레코드가 없습니다. 프로비저닝 시 자동 생성됩니다.",
-    "No jobs match this filter.": "이 필터에 해당하는 작업이 없습니다.",
-    "No pending or running jobs in the operator queue.": "오퍼레이터 큐에 대기 중이거나 실행 중인 작업이 없습니다.",
-    "No provider allocation data yet.": "아직 프로바이더 할당 데이터가 없습니다.",
-    "No recent failed jobs.": "최근 실패한 작업이 없습니다.",
-    "No recent superseded jobs.": "최근 대체된 작업이 없습니다.",
-    "No source groups match this filter.": "이 필터에 해당하는 소스 그룹이 없습니다.",
-    "No worlds match this filter.": "이 필터에 해당하는 월드가 없습니다.",
-    "Open world fleet": "월드 플릿 열기",
-    "Operator Queue": "오퍼레이터 큐",
-    "Ops note": "운영 메모",
-    "Pending and running lifecycle work across the instance fleet.": "인스턴스 플릿 전반의 대기 중 및 실행 중인 라이프사이클 작업입니다.",
-    "Plan code": "요금제 코드",
-    "Plan name": "요금제 이름",
-    "Previous page": "이전 페이지",
-    "Price (cents)": "가격(센트)",
-    "Private IP": "사설 IP",
-    "Project": "프로젝트",
-    "Provider instance": "프로바이더 인스턴스",
-    "Provider message": "프로바이더 메시지",
-    "Provider snapshot": "프로바이더 스냅샷",
-    "Provider spread": "프로바이더 분산",
-    "Provider volume": "프로바이더 볼륨",
-    "Public IP": "공인 IP",
-    "Publicly purchasable": "공개 구매 가능",
-    "Queuing replay...": "재실행 큐에 추가 중…",
-    "Raw status": "원시 상태",
-    "Recent Failures": "최근 실패",
-    "Refresh status": "상태 새로고침",
-    "Refreshing...": "새로고침 중…",
-    "Remote host": "원격 호스트",
-    "Remote path": "원격 경로",
-    "Replay matching failed tasks": "일치하는 실패 작업 재실행",
-    "Result": "결과",
-    "Revoke all matching": "일치하는 항목 모두 취소",
-    "Revoke group": "그룹 취소",
-    "Revoke matching risk groups": "일치하는 위험 그룹 취소",
-    "Revoke now": "즉시 취소",
-    "Revoke selected": "선택 항목 취소",
-    "Running now": "현재 실행 중",
-    "Runtime is not checking in within the configured stale window.": "런타임이 설정된 만료 기간 내에 보고하지 않고 있습니다.",
-    "Runtime observation": "런타임 관측",
-    "Save world": "월드 저장",
-    "Saving...": "저장 중…",
-    "Search worlds": "월드 검색",
-    "Select active sessions on this page to revoke them in one action.": "이 페이지의 활성 세션을 선택하면 한 번에 취소할 수 있습니다.",
-    "Select provider": "프로바이더 선택",
-    "Shown to users in the app subscription page.": "앱 구독 페이지에서 사용자에게 표시됩니다.",
-    "Snapshots": "스냅샷",
-    "Sort order": "정렬 순서",
-    "Summary counts reflect all jobs matching the current filters, not just this page.": "요약 수치는 현재 페이지가 아닌 현재 필터에 일치하는 모든 작업을 반영합니다.",
-    "Superseded Queue": "대체된 큐",
-    "Top attention items": "최우선 주의 항목",
-    "Total results": "총 결과",
-    "Trial": "체험",
-    "Value (JSON)": "값(JSON)",
-    "Viewing sessions": "세션 보는 중",
-    "Visible failed": "표시된 실패",
-    "Visible pending": "표시된 대기",
-    "Visible rows": "표시된 행",
-    "Visible running": "표시된 실행",
-    "World drift summary": "월드 드리프트 요약",
-    "World failed jobs": "월드 실패 작업",
-    "World name": "월드 이름",
-    "World operator jobs": "월드 오퍼레이터 작업",
-    "World scope": "월드 범위",
-    "World superseded jobs": "월드 대체된 작업",
-    "Worlds that currently need operator attention.": "현재 오퍼레이터의 주의가 필요한 월드입니다.",
-    "Yes": "예",
-    "e.g. New user trial": "예: 신규 사용자 체험",
-    "e.g. trial": "예: trial",
-    "{count} active session(s) selected on this page.": "이 페이지에서 활성 세션 {count}개 선택됨.",
-    "{count} jobs": "작업 {count}개",
-  },
-};
-
-const cloudConsoleChineseTextOverrides: Record<string, string> = {
-  Actions: "操作",
-  Access: "访问",
-  Action: "操作",
-  Active: "启用",
-  Admin: "管理端",
-  "Config key": "配置键",
-  "Internal note shown next to the key.": "在配置键旁边显示的内部备注。",
-  "Value (JSON)": "JSON 值",
-  "manual-ban": "手动封禁",
-  "Saving...": "保存中…",
-  "Save world": "保存世界",
-  "Refreshing...": "刷新中…",
-  "Refresh status": "刷新状态",
-  "Health message": "健康信息",
-  Name: "名称",
-  "Provider instance": "供应方实例",
-  "Provider volume": "供应方卷",
-  "Provider snapshot": "供应方快照",
-  "Private IP": "内网 IP",
-  "Public IP": "公网 IP",
-  Region: "区域",
-  Zone: "可用区",
-  Image: "镜像",
-  Flavor: "规格",
-  Disk: "磁盘",
-  Bootstrapped: "已启动",
-  Created: "创建时间",
-  Updated: "更新时间",
-  "Provider message": "供应方消息",
-  "Deployment state": "部署状态",
-  "Deployment mode": "部署模式",
-  "Executor mode": "执行模式",
-  "Remote host": "远端主机",
-  "Remote path": "远端路径",
-  Project: "项目",
-  Container: "容器",
-  "Raw status": "原始状态",
-  "No instance record exists yet. Provisioning will create one automatically.":
-    "尚未创建实例记录。下一次创建任务会自动建立。",
-  "Provider-side deployment status observed from the current compute adapter.":
-    "当前算力适配器观测到的供应方端部署状态。",
-  "Jobs show how this world moved through provision, resume, and suspend work.":
-    "任务记录该世界在创建、恢复、挂起等生命周期中的执行情况。",
-  "Plan code": "套餐代码",
-  "Plan name": "套餐名称",
-  "Duration (days)": "时长（天）",
-  "Price (cents)": "价格（分）",
-  Currency: "货币",
-  "Sort order": "排序",
-  Description: "说明",
-  Trial: "试用套餐",
-  "Publicly purchasable": "公开可购买",
-  "e.g. trial": "例如 trial",
-  "e.g. New user trial": "例如 新用户试用",
-  "Shown to users in the app subscription page.": "展示在 App 订阅页给用户看的文案。",
-  Code: "代码",
-  "Admin:": "管理端：",
-  API: "API",
-  "API:": "API：",
-  "Artifact summary": "工件摘要",
-  "Alert status": "告警状态",
-  "Attention worlds": "关注世界",
-  "attention:": "关注：",
-  "audit:": "审计：",
-  "active providers": "活跃供应方",
-  "Activity endpoint:": "活动端点：",
-  "Approval and manual delivery queue shortcuts wired to the request projection filters.":
-    "审批与手动交付队列快捷入口已接入申请投影筛选器。",
-  "Approval and projected-world shortcuts surfaced in the main dashboard header.":
-    "审批与预计世界状态快捷入口已显示在主仪表盘顶部。",
-  "Accounts and expiry": "账号与到期",
-  "Aggregate sessions by issue IP and client under the current filters, then revoke an entire source in one action.":
-    "按当前筛选条件聚合签发 IP 与客户端来源，并可一次性吊销整个来源。",
-  "Alerts upgraded because retry or stale-heartbeat thresholds were crossed.":
-    "由于重试次数或心跳陈旧阈值被触发，告警已升级。",
-  "Bootstrap endpoint:": "启动端点：",
-  "Bootstrapped:": "已启动：",
-  "Callback token rotated.": "回调令牌已轮换。",
-  "Callback token:": "回调令牌：",
-  "Clear context focus": "清除上下文聚焦",
-  "Clear matching failed tasks": "清除匹配的失败任务",
-  "Clear now": "立即清除",
-  "Clear receipts": "清除回执",
-  "Clear source focus": "清除来源聚焦",
-  "Clear task highlight": "清除任务高亮",
-  "Clear world scope": "清除世界范围",
-  "Close review": "关闭复核",
-  "Cloud platform:": "云平台：",
-  "Configure role usage pricing, payees, contribution weights, and settlement ledgers.":
-    "配置角色使用定价、收益人、贡献权重与结算账本。",
-  "Configure trial, monthly, quarterly, yearly, and invite reward plans shown to the app.":
-    "配置 App 展示的试用、月付、季付、年付和邀请奖励套餐。",
-  "Compatibility view for the older approval workflow. Staff can still use it when manual intervention is needed.":
-    "旧审批流程的兼容视图，需要人工介入时运维人员仍可使用。",
-  "Context groups": "上下文分组",
-  "Container:": "容器：",
-  "Copy compose": "复制 Compose",
-  "Copy endpoints": "复制端点",
-  "Copy env": "复制环境变量",
-  "Copy jobs permalink": "复制任务固定链接",
-  "Copy requests permalink": "复制申请固定链接",
-  "Copy review context": "复制复核上下文",
-  "Copy sessions permalink": "复制会话固定链接",
-  "Copy task context": "复制任务上下文",
-  "Copy token": "复制令牌",
-  "Copy waiting sync permalink": "复制等待同步固定链接",
-  "Copy worlds permalink": "复制世界固定链接",
-  "Critical alerts": "严重告警",
-  "Critical stale threshold:": "严重心跳陈旧阈值：",
-  "Critical threshold": "严重阈值",
-  "Current alert severity after applying retry and stale-heartbeat thresholds.":
-    "应用重试与心跳陈旧阈值后的当前告警严重级别。",
-  "Current endpoints": "当前端点",
-  "Current rationale": "当前依据",
-  "Current session": "当前会话",
-  "Daily summary": "每日摘要",
-  "Default region:": "默认区域：",
-  "Default zone:": "默认可用区：",
-  Delayed: "延迟",
-  Running: "运行中",
-  Other: "其他",
-  "Lease expired": "租约过期",
-  "{count} jobs": "{count} 项任务",
-  "Deployment mode:": "部署模式：",
-  "Deployment state:": "部署状态：",
-  "Deployment:": "部署：",
-  "Executor:": "执行模式：",
-  "Provider:": "供应方：",
-  "Derived from session issue, expiry, and revoke events inside the focused source group under the current filters.":
-    "基于当前筛选条件下聚焦来源分组内的会话签发、过期与吊销事件生成。",
-  "Desired state": "期望状态",
-  "Direction": "方向",
-  "Download CSV": "下载 CSV",
-  "Download JSON": "下载 JSON",
-  "Export context CSV": "导出上下文 CSV",
-  "Export context snapshot": "导出上下文快照",
-  "Export focus CSV": "导出聚焦 CSV",
-  "Export focus snapshot": "导出聚焦快照",
-  "Export matching CSV": "导出匹配 CSV",
-  "Export matching snapshot": "导出匹配快照",
-  "Export risk CSV": "导出风险 CSV",
-  "Export risk snapshot": "导出风险快照",
-  "Export snapshot": "导出快照",
-  Escalated: "已升级",
-  "Failure endpoint:": "失败端点：",
-  "Focused source": "聚焦来源",
-  "Focus path:": "聚焦路径：",
-  "Group risk": "分组风险",
-  "Manage SaaS trial, invite, feature, copy, and public app URL configuration.":
-    "管理 SaaS 试用、邀请、功能、文案和公开 App URL 配置。",
-  "Health endpoint:": "健康端点：",
-  "Heartbeat endpoint:": "心跳端点：",
-  "Heartbeat stale": "心跳陈旧",
-  "Image:": "镜像：",
-  Instance: "实例",
-  "Issued from": "签发来源",
-  "Job status": "任务状态",
-  "Last client": "最近客户端",
-  "Manual Docker": "手动 Docker",
-  "Managed lifecycle": "托管生命周期",
-  "Matching sessions": "匹配会话",
-  "No requests match this filter.": "没有申请匹配当前筛选。",
-  "No worlds match this filter.": "没有世界匹配当前筛选。",
-  "No jobs match this filter.": "没有任务匹配当前筛选。",
-  "No waiting sync tasks match this filter.":
-    "没有等待同步任务匹配当前筛选。",
-  "No admin sessions match this filter.": "没有管理会话匹配当前筛选。",
-  "No active attention items. The fleet currently looks healthy.":
-    "当前没有活跃关注项。舰队状态看起来健康。",
-  "No pending or running jobs in the operator queue.":
-    "运维队列暂无待处理或运行中的任务。",
-  "No provider allocation data yet.": "暂无供应方分配数据。",
-  "No recent failed jobs.": "暂无最近失败任务。",
-  "No recent superseded jobs.": "暂无最近已替代任务。",
-  "No request workflow items yet.": "暂无申请工作流条目。",
-  Next: "下一页",
-  "Ops note": "运维备注",
-  Page: "页",
-  "Payees and ledgers": "收益人与账本",
-  "Page size": "每页数量",
-  "per page": "每页",
-  Previous: "上一页",
-  "Pricing and access": "价格与访问",
-  "Project:": "项目：",
-  "Projected world": "预计世界",
-  "power:": "电源：",
-  "provider:": "供应方：",
-  "provider: all": "供应方：全部",
-  "queue:": "队列：",
-  "queue: all": "队列：全部",
-  "queue: running": "队列：运行中",
-  "queue: lease expired": "队列：租约过期",
-  "queue: delayed": "队列：延迟",
-  "Recent operation receipts": "最近操作回执",
-  "Remote host:": "远程主机：",
-  "Remote path:": "远程路径：",
-  "Replay now": "立即重放",
-  "Request Workflow": "申请工作流",
-  "Reset filters": "重置筛选",
-  "Result": "结果",
-  "Retry count": "重试次数",
-  "Retry count:": "重试次数：",
-  "Retry threshold:": "重试阈值：",
-  "Review live admin sessions, inspect where they were issued from, filter by revocation path, and page through longer audit history.":
-    "查看实时管理会话，检查签发来源，按吊销路径筛选，并分页浏览更长的审计历史。",
-  "Review cloud account status, subscription expiry, inviter attribution, and world linkage.":
-    "复核云账号状态、订阅到期、邀请人归因和世界关联。",
-  "Review reward redemptions, device and IP signals, and manual reversals.":
-    "复核奖励兑换、设备与 IP 信号以及人工撤销。",
-  "Review requests (": "复核申请（",
-  "Rewards and risk": "奖励与风控",
-  Reviewing: "复核中",
-  Revocation: "吊销",
-  "Revocation reason": "吊销原因",
-  "Revoked by session": "由会话吊销",
-  "Risk timeline": "风险时间线",
-  "Risk timeline data is not ready for export yet.":
-    "风险时间线数据尚未准备好，暂不能导出。",
-  "Runbook": "运行手册",
-  "Running now": "正在运行",
-  "Runtime env overlay": "运行时环境覆盖",
-  "Runtime is not checking in within the configured stale window.":
-    "运行时未在配置的陈旧窗口内上报心跳。",
-  "Runtime observation": "运行时观测",
-  Scope: "范围",
-  "Search requests": "搜索申请",
-  "Search worlds": "搜索世界",
-  "Select all active admin sessions": "选择所有活跃管理会话",
-  "Select provider": "选择供应方",
-  "Session context": "会话上下文",
-  Showing: "显示",
-  "Showing the latest": "显示最新",
-  "Snapshots:": "快照：",
-  "Sort by": "排序字段",
-  "sort by:": "排序：",
-  "direction:": "方向：",
-  "Source context": "来源上下文",
-  "Source direction": "来源排序方向",
-  "Source groups": "来源分组",
-  "Source page size": "来源每页数量",
-  "Source risk": "来源风险",
-  "Source sort": "来源排序",
-  "Stale heartbeat seconds:": "心跳陈旧秒数：",
-  "status:": "状态：",
-  "health:": "健康：",
-  "superseded by:": "替代来源：",
-  "Superseded Queue": "已替代队列",
-  "Suggested admin:": "建议管理地址：",
-  "Suggested API:": "建议 API：",
-  "Summary counts reflect all jobs matching the current filters, not just this page.":
-    "摘要计数反映所有匹配当前筛选的任务，而不只是当前页。",
-  "Synced with": "已同步到",
-  "Target values:": "目标值：",
-  Targets: "目标",
-  Task: "任务",
-  "Trial and copy": "试用与文案",
-  "Task highlighted": "任务已高亮",
-  "Task permalink focus": "任务固定链接聚焦",
-  "Task types:": "任务类型：",
-  "The world will move toward sleeping state and active sessions may need to reconnect after it wakes again.":
-    "世界将进入休眠状态，重新唤醒后活跃会话可能需要重新连接。",
-  "These sessions reflect the latest event captured in this summary.":
-    "这些会话反映本摘要捕获到的最新事件。",
-  "This panel folds together runtime heartbeat freshness, provider-observed drift, and queued recovery jobs.":
-    "此面板汇总运行时心跳新鲜度、供应方观测到的漂移以及已排队的恢复任务。",
-  "This provider key is not in the current catalog yet.":
-    "当前目录中还没有这个供应方 key。",
-  "This task is no longer visible in the current result set, but recent receipts still match this task permalink.":
-    "此任务已不在当前结果集中可见，但最近回执仍匹配该任务固定链接。",
-  "This will queue a new recovery action and clear the current failure state for the world.":
-    "这会排队一个新的恢复动作，并清除该世界当前的失败状态。",
-  "Timeline audit detail": "时间线审计详情",
-  "Timeline focus": "时间线聚焦",
-  Timing: "时间",
-  "Top attention items": "重点关注项",
-  "Total fleet:": "舰队总数：",
-  "Total results": "结果总数",
-  "type:": "类型：",
-  Unassigned: "未分配",
-  "Use this env overlay when deploying the user's dedicated world runtime.":
-    "部署用户专属世界运行时时使用这份环境变量覆盖。",
-  "View in sessions list": "在会话列表中查看",
-  "Viewing source group": "正在查看来源分组",
-  "Visible contexts from the current page. Use these cards to focus a single recovery path before replaying or clearing failed tasks.":
-    "当前页可见上下文。可先用这些卡片聚焦单条恢复路径，再重放或清除失败任务。",
-  "Visible failed": "可见失败",
-  "Visible pending": "可见待处理",
-  "Visible rows": "可见行",
-  "Visible running": "可见运行中",
-  "visible task(s) · latest update": "个可见任务 · 最近更新",
-  "visible task(s) from": "个可见任务，来自",
-  "Volume:": "卷：",
-  "Waiting sync task clear was skipped.": "等待同步任务清除已跳过。",
-  "Waiting sync task cleared.": "等待同步任务已清除。",
-  "Waiting sync task replay queued.": "等待同步任务重放已排队。",
-  "Waiting sync task replay was skipped.": "等待同步任务重放已跳过。",
-  "Watch threshold": "观察阈值",
-  "Weekly summary": "每周摘要",
-  "Work from the instance point of view: provider placement, power state, heartbeat freshness, and quick lifecycle actions.":
-    "从实例视角工作：供应方部署位置、电源状态、心跳新鲜度和快捷生命周期操作。",
-  "World detail:": "世界详情：",
-  "World drift summary": "世界漂移摘要",
-  "World failed jobs": "世界失败任务",
-  "World operator jobs": "世界运维任务",
-  "World request saved.": "世界申请已保存。",
-  "World scope": "世界范围",
-  "World settings saved.": "世界设置已保存。",
-  "World superseded jobs": "世界已替代任务",
-  "Worlds already in critical state, including failed and escalated alerts.":
-    "已处于严重状态的世界，包括失败与已升级告警。",
-  "Worlds that already have active `resume` or `provision` work in flight.":
-    "已有活跃恢复或创建任务进行中的世界。",
-  "Worlds that currently need operator attention.":
-    "当前需要运维关注的世界。",
-  "Zone:": "可用区：",
-
-  "Quick view of world availability, queued recovery, and the most urgent runtime drift signals across the cloud fleet.":
-    "快速查看云世界舰队的可用性、已排队恢复任务和最紧急的运行时漂移信号。",
-  "Inspect jobs": "检查任务",
-  "PROJECTED:": "预计：",
-  "DESIRED:": "期望：",
-  "Projected:": "预计：",
-  "Desired:": "期望：",
-  Creating: "创建中",
-  CREATING: "创建中",
-  Bootstrapping: "启动配置中",
-  BOOTSTRAPPING: "启动配置中",
-  Cancelled: "已取消",
-  QUEUED: "已排队",
-  STARTING: "启动中",
-  READY: "就绪",
-  SLEEPING: "休眠",
-  FAILED: "失败",
-  DISABLED: "已停用",
-  ALL: "全部",
-  all: "全部",
-  active: "活跃",
-  revoked: "已吊销",
-  total: "条",
-  expired: "已过期",
-  "refresh reuse": "刷新令牌复用",
-  pending: "待处理",
-  running: "运行中",
-  succeeded: "成功",
-  failed: "失败",
-  cancelled: "已取消",
-  queued: "已排队",
-  creating: "创建中",
-  bootstrapping: "启动配置中",
-  starting: "启动中",
-  ready: "就绪",
-  sleeping: "休眠",
-  disabled: "已停用",
-  absent: "无实例",
-  provisioning: "创建中",
-  stopped: "已停止",
-  stopping: "停止中",
-  error: "错误",
-  healthy: "健康",
-  critical: "严重",
-  warning: "警告",
-  info: "信息",
-  unhealthy: "异常",
-  unknown: "未知",
-  normal: "正常",
-  watch: "观察",
-  provision: "创建",
-  resume: "恢复",
-  suspend: "挂起",
-  reconcile: "对账",
-  superseded: "已替代",
-  refresh_world: "刷新世界",
-  refresh_phone: "刷新手机号",
-  invalidate_phone: "使手机号失效",
-  updatedAt: "更新时间",
-  createdAt: "创建时间",
-  availableAt: "可用时间",
-  startedAt: "开始时间",
-  finishedAt: "完成时间",
-  expiresAt: "刷新过期",
-  lastUsedAt: "最近使用",
-  revokedAt: "吊销时间",
-  activeSessions: "活跃会话",
-  totalSessions: "总会话",
-  latestLastUsedAt: "最近使用",
-  latestCreatedAt: "最近创建",
-  latestRevokedAt: "最近吊销",
-  desc: "降序",
-  asc: "升序",
-  current: "当前",
-  logout: "退出登录",
-  "manual-revocation": "手动吊销",
-  "refresh-token-reuse": "刷新令牌复用",
-  running_now: "正在运行",
-  lease_expired: "租约过期",
-  "Warning:": "警告：",
-  "Heartbeat stale:": "心跳陈旧：",
-  "Failed:": "失败：",
-  "Instance pool": "实例池",
-  "Instance fleet": "实例舰队",
-  worlds: "个世界",
-  WORLDS: "个世界",
-  provider: "供应方",
-  PROVIDER: "供应方",
-  Provider: "供应方",
-  "Provider drift": "供应方漂移",
-  "Provider spread": "供应方分布",
-  "Open requests (": "打开申请（",
-  "Open failed jobs (": "打开失败任务（",
-  "Open superseded jobs (": "打开已替代任务（",
-  "Open superseded queue (": "打开已替代队列（",
-  "Open world fleet": "打开世界舰队",
-  "Loading request workflow...": "正在加载申请工作流...",
-  "Operator Queue": "运维队列",
-  "Pending and running lifecycle work across the instance fleet.":
-    "实例舰队中待处理和运行中的生命周期工作。",
-  "Loading lifecycle jobs...": "正在加载生命周期任务...",
-  "Recent Failures": "最近失败",
-  "Latest failed lifecycle work that may need manual recovery.":
-    "最近可能需要人工恢复的失败生命周期工作。",
-  "Latest lifecycle jobs that were replaced by newer work.":
-    "最近被新任务替代的生命周期任务。",
-  "Loading superseded lifecycle jobs...": "正在加载已替代生命周期任务...",
-  "Most urgent worlds needing manual inspection or follow-up.":
-    "最需要人工检查或跟进的紧急世界。",
-  "Loading fleet dashboard...": "正在加载舰队仪表盘...",
-
-  "Loading requests...": "正在加载申请...",
-  "Loading worlds...": "正在加载世界...",
-  "Loading drift summary...": "正在加载漂移摘要...",
-  "Loading world drift summary...": "正在加载世界漂移摘要...",
-  "Loading instances...": "正在加载实例...",
-  "Loading jobs...": "正在加载任务...",
-  "World name": "世界名",
-  Lease: "租约",
-  "Provider reports power state that disagrees with desired world state.":
-    "供应方上报的电源状态与世界期望状态不一致。",
-  "Each phone owns exactly one world. New users provision a fresh instance, while returning users wake their previous one.":
-    "每个手机号只拥有一个世界。新用户会创建新实例，老用户会唤醒之前的实例。",
-  "NO INSTANCE": "无实例",
-  "No instance": "无实例",
-  "NEEDS ACTION": "需要处理",
-  "Needs action": "需要处理",
-  "power: all": "电源：全部",
-  "power: absent": "电源：无实例",
-  "power: provisioning": "电源：创建中",
-  "power: running": "电源：运行中",
-  "power: starting": "电源：启动中",
-  "power: stopped": "电源：已停止",
-  "power: stopping": "电源：停止中",
-  "power: error": "电源：错误",
-  "attention: all": "关注：全部",
-  "attention: healthy": "关注：健康",
-  "attention: critical": "关注：严重",
-  "attention: warning": "关注：警告",
-  "attention: info": "关注：信息",
-  "health: all": "健康：全部",
-  "health: healthy": "健康：健康",
-  "health: unhealthy": "健康：异常",
-  "health: unknown": "健康：未知",
-
-  "Inspect provisioning, resume, suspend, and reconcile work across the managed world fleet.":
-    "检查托管世界舰队中的创建、恢复、挂起和对账工作。",
-  "status: all": "状态：全部",
-  "status: pending": "状态：待处理",
-  "status: running": "状态：运行中",
-  "status: succeeded": "状态：成功",
-  "status: failed": "状态：失败",
-  "status: cancelled": "状态：已取消",
-  "type: all": "类型：全部",
-  "type: provision": "类型：创建",
-  "type: resume": "类型：恢复",
-  "type: suspend": "类型：挂起",
-  "type: reconcile": "类型：对账",
-  "audit: all": "审计：全部",
-  "audit: superseded": "审计：已替代",
-  "superseded by: all": "替代来源：全部",
-  "superseded by: provision": "替代来源：创建",
-  "superseded by: resume": "替代来源：恢复",
-  "superseded by: suspend": "替代来源：挂起",
-  "superseded by: reconcile": "替代来源：对账",
-  "sort by: updatedAt": "排序：更新时间",
-  "sort by: createdAt": "排序：创建时间",
-  "sort by: availableAt": "排序：可用时间",
-  "sort by: startedAt": "排序：开始时间",
-  "sort by: finishedAt": "排序：完成时间",
-  "direction: desc": "方向：降序",
-  "direction: asc": "方向：升序",
-  "page size: 20": "每页：20",
-  "page size: 50": "每页：50",
-  "page size: 100": "每页：100",
-  "Previous page": "上一页",
-  "Next page": "下一页",
-  "Showing ": "显示 ",
-  " of ": " / ",
-  " jobs.": " 个任务。",
-
-  "SHOWING 0 SESSIONS": "正在显示 0 个会话",
-  "Showing 0 sessions": "正在显示 0 个会话",
-  "SHOWING 0 GROUPS": "正在显示 0 个分组",
-  "Showing 0 groups": "正在显示 0 个分组",
-  "QUICK VIEWS": "快捷视图",
-  "Quick views": "快捷视图",
-  "RECENTLY REVOKED": "最近吊销",
-  "Recently revoked": "最近吊销",
-  "EXPIRING SOON": "即将过期",
-  "Expiring soon": "即将过期",
-  "Loading source groups...": "正在加载来源分组...",
-  "Loading admin sessions...": "正在加载管理会话...",
-  "REVOKE ALL MATCHING": "吊销所有匹配项",
-  "Revoke all matching": "吊销所有匹配项",
-  "Manual revoke": "手动吊销",
-  "Refresh reuse": "刷新复用",
-  "Refresh expiry": "刷新过期",
-  "Active sessions": "活跃会话",
-  "Total sessions": "总会话",
-  "Latest used": "最近使用",
-  "Latest created": "最近创建",
-  "Latest revoked": "最近吊销",
-  "All risk levels": "全部风险级别",
-  "Critical only": "仅严重",
-  "Watch only": "仅观察",
-  "Normal only": "仅正常",
-  "ALL RISK": "全部风险",
-  "All risk": "全部风险",
-  "CRITICAL RISK": "严重风险",
-  "Critical risk": "严重风险",
-  "WATCH RISK": "观察风险",
-  "Watch risk": "观察风险",
-  "NORMAL RISK": "正常风险",
-  "Normal risk": "正常风险",
-  "Select a source risk filter to batch-revoke risky groups.":
-    "选择来源风险筛选后，可批量吊销风险分组。",
-  "EXPORT RISK GROUPS CSV": "导出风险分组 CSV",
-  "Export risk groups CSV": "导出风险分组 CSV",
-
-  "DURABLE RECOVERY": "持久化恢复",
-  "Durable recovery": "持久化恢复",
-  "Inspect durable waiting-session compensation tasks, replay exhausted retries, and clear stale failures without leaving the console.":
-    "在控制台内检查持久化等待会话补偿任务，重放已耗尽重试的任务，并清理陈旧失败。",
-  "All task types": "全部任务类型",
-  "Replay matching failed tasks": "重放匹配的失败任务",
-  "Queuing replay...": "正在加入重放队列…",
-  "Resume is available for worlds that still need to move back toward running, including sleeping, failed, queued, and stopping states. Suspend is limited to worlds that are currently active. Retry is reserved for failed or in-flight lifecycle states.":
-    "适用于还需回到运行状态的世界（包含休眠、失败、排队、停止中），即可执行恢复。挂起仅作用于当前活跃的世界，重试仅适用于失败或运行中的生命周期状态。",
-  Deployment: "部署模式",
-  "Default region": "默认区域",
-  "Default zone": "默认可用区",
-  "Managed provisioning": "托管供应",
-  Snapshots: "快照",
-  Yes: "是",
-  No: "否",
-  "Export risk sessions CSV": "导出风险会话 CSV",
-  "Revoke matching risk groups": "吊销匹配的风险来源",
-  "Revoke group": "吊销该来源",
-  "Revoke now": "立即吊销",
-  "Revoke selected": "吊销所选会话",
-  "Viewing sessions": "正在查看会话",
-  "View sessions": "查看会话",
-  "No source groups match this filter.": "没有来源分组匹配当前筛选。",
-  "Select active sessions on this page to revoke them in one action.":
-    "勾选本页中的活跃会话，可一次性批量吊销。",
-  "{count} active session(s) selected on this page.":
-    "本页已选中 {count} 个活跃会话。",
-  "Export filtered snapshot": "导出筛选快照",
-  "Export filtered CSV": "导出筛选 CSV",
-  "Export context groups snapshot": "导出上下文分组快照",
-  "Export context groups CSV": "导出上下文分组 CSV",
-  "Open waiting sync permalink": "打开等待同步固定链接",
-  "No visible contexts on this page yet.": "当前页还没有可见上下文。",
-  "Loading waiting sync tasks...": "正在加载等待同步任务...",
-  "页 1 of 1": "第 1 页 / 共 1 页",
-
-  // Bootstrap config / alert / jobs queue / pagination labels missing from
-  // the rest of the dict.
-  "Fail endpoint:": "失败端点：",
-  "Escalated:": "已升级：",
-  "Escalation reason:": "升级原因：",
-  "Not stale": "未陈旧",
-  "Queue totals reflect all jobs for this world, not just the recent 20 jobs below.":
-    "队列汇总反映该世界的全部任务，下方只展示最近 20 项。",
-  "Rotating...": "轮换中…",
-  "Rotate callback token": "轮换回调令牌",
-  "Phone unavailable": "无手机号",
-  "No tasks match this filter.": "没有任务匹配当前筛选。",
-  "Previous groups": "上一页分组",
-  "Next groups": "下一页分组",
-  "Event view": "事件视图",
-  "Current snapshot": "当前快照",
-  "Session issued": "会话签发",
-  "Session expired": "会话过期",
-  "Session revoked": "会话吊销",
-  "Unknown client": "未知客户端",
-  "Revoke session": "吊销该会话",
-  "Open user": "打开用户",
-  remaining: "剩余",
-  expires: "到期",
-  available: "可用",
-};
-
-const cloudConsoleJapaneseTextOverrides: Record<string, string> = {
-  Actions: "操作",
-  Access: "アクセス",
-  Action: "操作",
-  Active: "有効",
-  Admin: "管理",
-  "Config key": "設定キー",
-  "Internal note shown next to the key.": "キーの隣に表示する内部メモ。",
-  "Value (JSON)": "値 (JSON)",
-  "manual-ban": "手動禁止",
-  "Saving...": "保存中…",
-  "Save world": "ワールドを保存",
-  "Refreshing...": "更新中…",
-  "Refresh status": "ステータスを更新",
-  "Health message": "ヘルスメッセージ",
-  Name: "名前",
-  "Provider instance": "プロバイダーインスタンス",
-  "Provider volume": "プロバイダーボリューム",
-  "Provider snapshot": "プロバイダースナップショット",
-  "Private IP": "プライベート IP",
-  "Public IP": "パブリック IP",
-  Region: "リージョン",
-  Zone: "ゾーン",
-  Image: "イメージ",
-  Flavor: "フレーバー",
-  Disk: "ディスク",
-  Bootstrapped: "ブートストラップ完了",
-  Created: "作成日時",
-  Updated: "更新日時",
-  "Provider message": "プロバイダーメッセージ",
-  "Deployment state": "デプロイ状態",
-  "Deployment mode": "デプロイモード",
-  "Executor mode": "実行モード",
-  "Remote host": "リモートホスト",
-  "Remote path": "リモートパス",
-  Project: "プロジェクト",
-  Container: "コンテナ",
-  "Raw status": "生ステータス",
-  "No instance record exists yet. Provisioning will create one automatically.":
-    "インスタンスレコードはまだ存在しません。次回のプロビジョニングで自動的に作成されます。",
-  "Provider-side deployment status observed from the current compute adapter.":
-    "現在のコンピュートアダプターから観測したプロバイダー側のデプロイ状態。",
-  "Jobs show how this world moved through provision, resume, and suspend work.":
-    "ジョブはこのワールドのプロビジョニング、再開、停止などのライフサイクル状況を示します。",
-  "Plan code": "プランコード",
-  "Plan name": "プラン名",
-  "Duration (days)": "期間 (日)",
-  "Price (cents)": "価格 (セント)",
-  Currency: "通貨",
-  "Sort order": "並び順",
-  Description: "説明",
-  Trial: "トライアル",
-  "Publicly purchasable": "一般購入可能",
-  "e.g. trial": "例: trial",
-  "e.g. New user trial": "例: 新規ユーザートライアル",
-  "Shown to users in the app subscription page.": "アプリのサブスクリプションページでユーザーに表示される文言。",
-  Code: "コード",
-  "Admin:": "管理:",
-  API: "API",
-  "API:": "API:",
-  "Artifact summary": "アーティファクト概要",
-  "Alert status": "アラート状態",
-  "Attention worlds": "要注意ワールド",
-  "attention:": "注目:",
-  "audit:": "監査:",
-  "active providers": "アクティブプロバイダー",
-  "Activity endpoint:": "アクティビティエンドポイント:",
-  "Approval and manual delivery queue shortcuts wired to the request projection filters.":
-    "承認と手動引き渡しキューのショートカットがリクエストプロジェクションフィルターに接続されています。",
-  "Approval and projected-world shortcuts surfaced in the main dashboard header.":
-    "承認と予定ワールドのショートカットがメインダッシュボードのヘッダーに表示されています。",
-  "Accounts and expiry": "アカウントと有効期限",
-  "Aggregate sessions by issue IP and client under the current filters, then revoke an entire source in one action.":
-    "現在のフィルターで発行 IP とクライアント別にセッションを集計し、ソース全体を一括で取り消せます。",
-  "Alerts upgraded because retry or stale-heartbeat thresholds were crossed.":
-    "リトライまたはハートビート古さのしきい値超過によりアラートが格上げされました。",
-  "Bootstrap endpoint:": "ブートストラップエンドポイント:",
-  "Bootstrapped:": "ブートストラップ完了:",
-  "Callback token rotated.": "コールバックトークンをローテーションしました。",
-  "Callback token:": "コールバックトークン:",
-  "Clear context focus": "コンテキストフォーカスをクリア",
-  "Clear matching failed tasks": "一致する失敗タスクをクリア",
-  "Clear now": "今すぐクリア",
-  "Clear receipts": "受領をクリア",
-  "Clear source focus": "ソースフォーカスをクリア",
-  "Clear task highlight": "タスクハイライトをクリア",
-  "Clear world scope": "ワールドスコープをクリア",
-  "Close review": "レビューを閉じる",
-  "Cloud platform:": "クラウドプラットフォーム:",
-  "Configure role usage pricing, payees, contribution weights, and settlement ledgers.":
-    "ロール利用料金、受取人、貢献度の重み、精算台帳を設定します。",
-  "Configure trial, monthly, quarterly, yearly, and invite reward plans shown to the app.":
-    "アプリに表示されるトライアル、月額、四半期、年額、招待報酬プランを設定します。",
-  "Compatibility view for the older approval workflow. Staff can still use it when manual intervention is needed.":
-    "旧承認ワークフローの互換ビュー。手動介入が必要な場合は引き続き使用できます。",
-  "Context groups": "コンテキストグループ",
-  "Container:": "コンテナ:",
-  "Copy compose": "Compose をコピー",
-  "Copy endpoints": "エンドポイントをコピー",
-  "Copy env": "環境変数をコピー",
-  "Copy jobs permalink": "ジョブのパーマリンクをコピー",
-  "Copy requests permalink": "リクエストのパーマリンクをコピー",
-  "Copy review context": "レビューコンテキストをコピー",
-  "Copy sessions permalink": "セッションのパーマリンクをコピー",
-  "Copy task context": "タスクコンテキストをコピー",
-  "Copy token": "トークンをコピー",
-  "Copy waiting sync permalink": "待機同期のパーマリンクをコピー",
-  "Copy worlds permalink": "ワールドのパーマリンクをコピー",
-  "Critical alerts": "重大アラート",
-  "Critical stale threshold:": "重大ハートビート古さしきい値:",
-  "Critical threshold": "重大しきい値",
-  "Current alert severity after applying retry and stale-heartbeat thresholds.":
-    "リトライとハートビート古さしきい値適用後の現在のアラート重大度。",
-  "Current endpoints": "現在のエンドポイント",
-  "Current rationale": "現在の根拠",
-  "Current session": "現在のセッション",
-  "Daily summary": "日次サマリー",
-  "Default region:": "デフォルトリージョン:",
-  "Default zone:": "デフォルトゾーン:",
-  Delayed: "遅延",
-  Running: "実行中",
-  Other: "その他",
-  "Lease expired": "リース期限切れ",
-  "{count} jobs": "{count} 件のジョブ",
-  "Deployment mode:": "デプロイモード:",
-  "Deployment state:": "デプロイ状態:",
-  "Deployment:": "デプロイ:",
-  "Derived from session issue, expiry, and revoke events inside the focused source group under the current filters.":
-    "現在のフィルターで対象とするソースグループ内のセッション発行、期限切れ、取り消しイベントから生成されます。",
-  "Desired state": "期待状態",
-  Direction: "方向",
-  "Download CSV": "CSV をダウンロード",
-  "Download JSON": "JSON をダウンロード",
-  "Export context CSV": "コンテキスト CSV をエクスポート",
-  "Export context snapshot": "コンテキストスナップショットをエクスポート",
-  "Export focus CSV": "フォーカス CSV をエクスポート",
-  "Export focus snapshot": "フォーカススナップショットをエクスポート",
-  "Export matching CSV": "一致 CSV をエクスポート",
-  "Export matching snapshot": "一致スナップショットをエクスポート",
-  "Export risk CSV": "リスク CSV をエクスポート",
-  "Export risk snapshot": "リスクスナップショットをエクスポート",
-  "Export snapshot": "スナップショットをエクスポート",
-  Escalated: "格上げ済み",
-  "Failure endpoint:": "失敗エンドポイント:",
-  "Focused source": "フォーカス中のソース",
-  "Focus path:": "フォーカスパス:",
-  "Group risk": "グループリスク",
-  "Manage SaaS trial, invite, feature, copy, and public app URL configuration.":
-    "SaaS のトライアル、招待、機能、文言、公開アプリ URL の設定を管理します。",
-  "Health endpoint:": "ヘルスエンドポイント:",
-  "Heartbeat endpoint:": "ハートビートエンドポイント:",
-  "Heartbeat stale": "ハートビート古い",
-  "Image:": "イメージ:",
-  Instance: "インスタンス",
-  "Issued from": "発行元",
-  "Job status": "ジョブ状態",
-  "Last client": "最終クライアント",
-  "Manual Docker": "手動 Docker",
-  "Managed lifecycle": "マネージドライフサイクル",
-  "Matching sessions": "一致セッション",
-  "No requests match this filter.": "このフィルターに一致するリクエストはありません。",
-  "No worlds match this filter.": "このフィルターに一致するワールドはありません。",
-  "No jobs match this filter.": "このフィルターに一致するジョブはありません。",
-  "No waiting sync tasks match this filter.": "このフィルターに一致する待機同期タスクはありません。",
-  "No admin sessions match this filter.": "このフィルターに一致する管理セッションはありません。",
-  "No active attention items. The fleet currently looks healthy.":
-    "アクティブな注目項目はありません。フリートは現在健全に見えます。",
-  "No pending or running jobs in the operator queue.": "オペレーターキューに保留中または実行中のジョブはありません。",
-  "No provider allocation data yet.": "プロバイダー割り当てデータはまだありません。",
-  "No recent failed jobs.": "最近の失敗ジョブはありません。",
-  "No recent superseded jobs.": "最近の置換ジョブはありません。",
-  "No request workflow items yet.": "リクエストワークフロー項目はまだありません。",
-  Next: "次へ",
-  "Ops note": "運用メモ",
-  Page: "ページ",
-  "Payees and ledgers": "受取人と台帳",
-  "Page size": "ページサイズ",
-  "per page": "件/ページ",
-  Previous: "前へ",
-  "Pricing and access": "価格とアクセス",
-  "Project:": "プロジェクト:",
-  "Projected world": "予定ワールド",
-  "power:": "電源:",
-  "provider:": "プロバイダー:",
-  "provider: all": "プロバイダー: すべて",
-  "queue:": "キュー:",
-  "queue: all": "キュー: すべて",
-  "queue: running": "キュー: 実行中",
-  "queue: lease expired": "キュー: リース期限切れ",
-  "queue: delayed": "キュー: 遅延",
-  "Recent operation receipts": "最近の操作受領",
-  "Remote host:": "リモートホスト:",
-  "Remote path:": "リモートパス:",
-  "Replay now": "今すぐ再実行",
-  "Request Workflow": "リクエストワークフロー",
-  "Reset filters": "フィルターをリセット",
-  Result: "結果",
-  "Retry count": "リトライ回数",
-  "Retry count:": "リトライ回数:",
-  "Retry threshold:": "リトライしきい値:",
-  "Review live admin sessions, inspect where they were issued from, filter by revocation path, and page through longer audit history.":
-    "ライブ管理セッションを確認し、発行元を調査し、取り消し経路でフィルターし、長い監査履歴をページ送りします。",
-  "Review cloud account status, subscription expiry, inviter attribution, and world linkage.":
-    "クラウドアカウントの状態、サブスクリプション期限、招待者帰属、ワールド連携を確認します。",
-  "Review reward redemptions, device and IP signals, and manual reversals.":
-    "報酬の引き換え、デバイスと IP のシグナル、手動取り消しを確認します。",
-  "Review requests (": "リクエストをレビュー (",
-  "Rewards and risk": "報酬とリスク",
-  Reviewing: "レビュー中",
-  Revocation: "取り消し",
-  "Revocation reason": "取り消し理由",
-  "Revoked by session": "セッションによる取り消し",
-  "Risk timeline": "リスクタイムライン",
-  "Risk timeline data is not ready for export yet.": "リスクタイムラインのデータはまだエクスポートできません。",
-  Runbook: "ランブック",
-  "Running now": "実行中",
-  "Runtime env overlay": "ランタイム環境オーバーレイ",
-  "Runtime is not checking in within the configured stale window.":
-    "ランタイムが設定された古さウィンドウ内にチェックインしていません。",
-  "Runtime observation": "ランタイム観測",
-  Scope: "スコープ",
-  "Search requests": "リクエストを検索",
-  "Search worlds": "ワールドを検索",
-  "Select all active admin sessions": "すべてのアクティブな管理セッションを選択",
-  "Select provider": "プロバイダーを選択",
-  "Session context": "セッションコンテキスト",
-  Showing: "表示中",
-  "Showing the latest": "最新を表示中",
-  "Snapshots:": "スナップショット:",
-  "Sort by": "並び替え",
-  "sort by:": "並び替え:",
-  "direction:": "方向:",
-  "Source context": "ソースコンテキスト",
-  "Source direction": "ソース並び方向",
-  "Source groups": "ソースグループ",
-  "Source page size": "ソースページサイズ",
-  "Source risk": "ソースリスク",
-  "Source sort": "ソース並び替え",
-  "Stale heartbeat seconds:": "ハートビート古さ秒数:",
-  "status:": "ステータス:",
-  "health:": "ヘルス:",
-  "superseded by:": "置換元:",
-  "Superseded Queue": "置換キュー",
-  "Suggested admin:": "推奨管理アドレス:",
-  "Suggested API:": "推奨 API:",
-  "Summary counts reflect all jobs matching the current filters, not just this page.":
-    "サマリーカウントは現在のフィルターに一致するすべてのジョブを反映しており、このページだけではありません。",
-  "Synced with": "同期先",
-  "Target values:": "目標値:",
-  Targets: "ターゲット",
-  Task: "タスク",
-  "Trial and copy": "トライアルと文言",
-  "Task highlighted": "タスクをハイライト",
-  "Task permalink focus": "タスクのパーマリンクフォーカス",
-  "Task types:": "タスクタイプ:",
-  "The world will move toward sleeping state and active sessions may need to reconnect after it wakes again.":
-    "ワールドはスリープ状態に移行し、再起動後にアクティブセッションが再接続を必要とする場合があります。",
-  "These sessions reflect the latest event captured in this summary.":
-    "これらのセッションはこのサマリーで取得した最新イベントを反映しています。",
-  "This panel folds together runtime heartbeat freshness, provider-observed drift, and queued recovery jobs.":
-    "このパネルはランタイムのハートビート鮮度、プロバイダー観測のドリフト、キュー内のリカバリージョブをまとめて表示します。",
-  "This provider key is not in the current catalog yet.": "このプロバイダーキーは現在のカタログにまだ存在しません。",
-  "This task is no longer visible in the current result set, but recent receipts still match this task permalink.":
-    "このタスクは現在の結果セットには表示されませんが、最近の受領はこのタスクパーマリンクに一致します。",
-  "This will queue a new recovery action and clear the current failure state for the world.":
-    "これは新しいリカバリーアクションをキューに追加し、このワールドの現在の失敗状態をクリアします。",
-  "Timeline audit detail": "タイムライン監査詳細",
-  "Timeline focus": "タイムラインフォーカス",
-  Timing: "時刻",
-  "Top attention items": "重要注目項目",
-  "Total fleet:": "総フリート:",
-  "Total results": "総結果数",
-  "type:": "タイプ:",
-  Unassigned: "未割り当て",
-  "Use this env overlay when deploying the user's dedicated world runtime.":
-    "ユーザー専用ワールドランタイムをデプロイする際にこの環境変数オーバーレイを使用してください。",
-  "View in sessions list": "セッション一覧で表示",
-  "Viewing source group": "ソースグループを表示中",
-  "Visible contexts from the current page. Use these cards to focus a single recovery path before replaying or clearing failed tasks.":
-    "現在のページに表示されているコンテキストです。失敗タスクを再実行またはクリアする前に、これらのカードで単一のリカバリーパスにフォーカスできます。",
-  "Visible failed": "表示中の失敗",
-  "Visible pending": "表示中の保留",
-  "Visible rows": "表示中の行",
-  "Visible running": "表示中の実行中",
-  "visible task(s) · latest update": "件の表示タスク · 最新更新",
-  "visible task(s) from": "件の表示タスク、出元",
-  "Volume:": "ボリューム:",
-  "Waiting sync task clear was skipped.": "待機同期タスクのクリアはスキップされました。",
-  "Waiting sync task cleared.": "待機同期タスクをクリアしました。",
-  "Waiting sync task replay queued.": "待機同期タスクの再実行をキューに追加しました。",
-  "Waiting sync task replay was skipped.": "待機同期タスクの再実行はスキップされました。",
-  "Watch threshold": "監視しきい値",
-  "Weekly summary": "週次サマリー",
-  "Work from the instance point of view: provider placement, power state, heartbeat freshness, and quick lifecycle actions.":
-    "インスタンス視点での作業：プロバイダー配置、電源状態、ハートビート鮮度、クイックライフサイクル操作。",
-  "World detail:": "ワールド詳細:",
-  "World drift summary": "ワールドドリフトサマリー",
-  "World failed jobs": "ワールド失敗ジョブ",
-  "World operator jobs": "ワールド運用ジョブ",
-  "World request saved.": "ワールドリクエストを保存しました。",
-  "World scope": "ワールドスコープ",
-  "World settings saved.": "ワールド設定を保存しました。",
-  "World superseded jobs": "ワールド置換ジョブ",
-  "Worlds already in critical state, including failed and escalated alerts.":
-    "重大状態のワールド（失敗および格上げ済みアラートを含む）。",
-  "Worlds that already have active `resume` or `provision` work in flight.":
-    "既にアクティブな `resume` または `provision` 作業が進行中のワールド。",
-  "Worlds that currently need operator attention.": "現在オペレーターの注意が必要なワールド。",
-  "Zone:": "ゾーン:",
-  "Quick view of world availability, queued recovery, and the most urgent runtime drift signals across the cloud fleet.":
-    "クラウドフリート全体のワールド可用性、キュー中のリカバリー、最も緊急なランタイムドリフトシグナルをクイックビュー。",
-  "Inspect jobs": "ジョブを確認",
-  "PROJECTED:": "予定:",
-  "DESIRED:": "期待:",
-  "Projected:": "予定:",
-  "Desired:": "期待:",
-  Creating: "作成中",
-  CREATING: "作成中",
-  Bootstrapping: "ブートストラップ中",
-  BOOTSTRAPPING: "ブートストラップ中",
-  Cancelled: "キャンセル",
-  QUEUED: "キュー済み",
-  STARTING: "起動中",
-  READY: "準備完了",
-  SLEEPING: "スリープ",
-  FAILED: "失敗",
-  DISABLED: "無効",
-  ALL: "すべて",
-  all: "すべて",
-  active: "アクティブ",
-  revoked: "取り消し済み",
-  total: "件",
-  expired: "期限切れ",
-  "refresh reuse": "リフレッシュトークン再利用",
-  pending: "保留",
-  running: "実行中",
-  succeeded: "成功",
-  failed: "失敗",
-  cancelled: "キャンセル",
-  queued: "キュー済み",
-  creating: "作成中",
-  bootstrapping: "ブートストラップ中",
-  starting: "起動中",
-  ready: "準備完了",
-  sleeping: "スリープ",
-  disabled: "無効",
-  absent: "インスタンスなし",
-  provisioning: "プロビジョニング中",
-  stopped: "停止",
-  stopping: "停止中",
-  error: "エラー",
-  healthy: "正常",
-  critical: "重大",
-  warning: "警告",
-  info: "情報",
-  unhealthy: "異常",
-  unknown: "不明",
-  normal: "通常",
-  watch: "監視",
-  provision: "プロビジョン",
-  resume: "再開",
-  suspend: "停止",
-  reconcile: "整合",
-  superseded: "置換済み",
-  refresh_world: "ワールドを更新",
-  refresh_phone: "電話番号を更新",
-  invalidate_phone: "電話番号を無効化",
-  updatedAt: "更新日時",
-  createdAt: "作成日時",
-  availableAt: "利用可能日時",
-  startedAt: "開始日時",
-  finishedAt: "完了日時",
-  expiresAt: "リフレッシュ期限",
-  lastUsedAt: "最終使用",
-  revokedAt: "取り消し日時",
-  activeSessions: "アクティブセッション",
-  totalSessions: "総セッション",
-  latestLastUsedAt: "最終使用",
-  latestCreatedAt: "最終作成",
-  latestRevokedAt: "最終取り消し",
-  desc: "降順",
-  asc: "昇順",
-  current: "現在",
-  logout: "ログアウト",
-  "manual-revocation": "手動取り消し",
-  "refresh-token-reuse": "リフレッシュトークン再利用",
-  running_now: "実行中",
-  lease_expired: "リース期限切れ",
-  "Warning:": "警告:",
-  "Heartbeat stale:": "ハートビート古い:",
-  "Failed:": "失敗:",
-  "Instance pool": "インスタンスプール",
-  "Instance fleet": "インスタンスフリート",
-  worlds: "件のワールド",
-  WORLDS: "件のワールド",
-  provider: "プロバイダー",
-  PROVIDER: "プロバイダー",
-  Provider: "プロバイダー",
-  "Provider drift": "プロバイダードリフト",
-  "Provider spread": "プロバイダー分布",
-  "Open requests (": "リクエストを開く (",
-  "Open failed jobs (": "失敗ジョブを開く (",
-  "Open superseded jobs (": "置換ジョブを開く (",
-  "Open superseded queue (": "置換キューを開く (",
-  "Open world fleet": "ワールドフリートを開く",
-  "Loading request workflow...": "リクエストワークフローを読み込み中...",
-  "Operator Queue": "オペレーターキュー",
-  "Pending and running lifecycle work across the instance fleet.":
-    "インスタンスフリート全体の保留中および実行中のライフサイクル作業。",
-  "Loading lifecycle jobs...": "ライフサイクルジョブを読み込み中...",
-  "Recent Failures": "最近の失敗",
-  "Latest failed lifecycle work that may need manual recovery.":
-    "手動リカバリーが必要になる可能性がある最新の失敗ライフサイクル作業。",
-  "Latest lifecycle jobs that were replaced by newer work.": "新しい作業に置き換えられた最新のライフサイクルジョブ。",
-  "Loading superseded lifecycle jobs...": "置換ライフサイクルジョブを読み込み中...",
-  "Most urgent worlds needing manual inspection or follow-up.": "手動検査またはフォローアップが最も必要な緊急ワールド。",
-  "Loading fleet dashboard...": "フリートダッシュボードを読み込み中...",
-  "Loading requests...": "リクエストを読み込み中...",
-  "Loading worlds...": "ワールドを読み込み中...",
-  "Loading drift summary...": "ドリフトサマリーを読み込み中...",
-  "Loading world drift summary...": "ワールドドリフトサマリーを読み込み中...",
-  "Loading instances...": "インスタンスを読み込み中...",
-  "Loading jobs...": "ジョブを読み込み中...",
-  "World name": "ワールド名",
-  Lease: "リース",
-  "Provider reports power state that disagrees with desired world state.":
-    "プロバイダーが報告する電源状態がワールドの期待状態と一致していません。",
-  "Each phone owns exactly one world. New users provision a fresh instance, while returning users wake their previous one.":
-    "各電話番号は 1 つのワールドのみを所有します。新規ユーザーは新しいインスタンスをプロビジョニングし、再訪ユーザーは以前のインスタンスを起こします。",
-  "NO INSTANCE": "インスタンスなし",
-  "No instance": "インスタンスなし",
-  "NEEDS ACTION": "対応が必要",
-  "Needs action": "対応が必要",
-  "power: all": "電源: すべて",
-  "power: absent": "電源: インスタンスなし",
-  "power: provisioning": "電源: プロビジョニング中",
-  "power: running": "電源: 実行中",
-  "power: starting": "電源: 起動中",
-  "power: stopped": "電源: 停止",
-  "power: stopping": "電源: 停止中",
-  "power: error": "電源: エラー",
-  "attention: all": "注目: すべて",
-  "attention: healthy": "注目: 正常",
-  "attention: critical": "注目: 重大",
-  "attention: warning": "注目: 警告",
-  "attention: info": "注目: 情報",
-  "health: all": "ヘルス: すべて",
-  "health: healthy": "ヘルス: 正常",
-  "health: unhealthy": "ヘルス: 異常",
-  "health: unknown": "ヘルス: 不明",
-  "Inspect provisioning, resume, suspend, and reconcile work across the managed world fleet.":
-    "マネージドワールドフリート全体のプロビジョニング、再開、停止、整合作業を確認します。",
-  "status: all": "ステータス: すべて",
-  "status: pending": "ステータス: 保留",
-  "status: running": "ステータス: 実行中",
-  "status: succeeded": "ステータス: 成功",
-  "status: failed": "ステータス: 失敗",
-  "status: cancelled": "ステータス: キャンセル",
-  "type: all": "タイプ: すべて",
-  "type: provision": "タイプ: プロビジョン",
-  "type: resume": "タイプ: 再開",
-  "type: suspend": "タイプ: 停止",
-  "type: reconcile": "タイプ: 整合",
-  "audit: all": "監査: すべて",
-  "audit: superseded": "監査: 置換済み",
-  "superseded by: all": "置換元: すべて",
-  "superseded by: provision": "置換元: プロビジョン",
-  "superseded by: resume": "置換元: 再開",
-  "superseded by: suspend": "置換元: 停止",
-  "superseded by: reconcile": "置換元: 整合",
-  "sort by: updatedAt": "並び替え: 更新日時",
-  "sort by: createdAt": "並び替え: 作成日時",
-  "sort by: availableAt": "並び替え: 利用可能日時",
-  "sort by: startedAt": "並び替え: 開始日時",
-  "sort by: finishedAt": "並び替え: 完了日時",
-  "direction: desc": "方向: 降順",
-  "direction: asc": "方向: 昇順",
-  "page size: 20": "ページサイズ: 20",
-  "page size: 50": "ページサイズ: 50",
-  "page size: 100": "ページサイズ: 100",
-  "Previous page": "前のページ",
-  "Next page": "次のページ",
-  "Showing ": "表示中 ",
-  " of ": " / ",
-  " jobs.": " 件のジョブ。",
-  "SHOWING 0 SESSIONS": "0 件のセッションを表示中",
-  "Showing 0 sessions": "0 件のセッションを表示中",
-  "SHOWING 0 GROUPS": "0 件のグループを表示中",
-  "Showing 0 groups": "0 件のグループを表示中",
-  "QUICK VIEWS": "クイックビュー",
-  "Quick views": "クイックビュー",
-  "RECENTLY REVOKED": "最近の取り消し",
-  "Recently revoked": "最近の取り消し",
-  "EXPIRING SOON": "まもなく期限切れ",
-  "Expiring soon": "まもなく期限切れ",
-  "Loading source groups...": "ソースグループを読み込み中...",
-  "Loading admin sessions...": "管理セッションを読み込み中...",
-  "REVOKE ALL MATCHING": "一致するものをすべて取り消し",
-  "Revoke all matching": "一致するものをすべて取り消し",
-  "Manual revoke": "手動取り消し",
-  "Refresh reuse": "リフレッシュ再利用",
-  "Refresh expiry": "リフレッシュ期限",
-  "Active sessions": "アクティブセッション",
-  "Total sessions": "総セッション",
-  "Latest used": "最終使用",
-  "Latest created": "最終作成",
-  "Latest revoked": "最終取り消し",
-  "All risk levels": "すべてのリスクレベル",
-  "Critical only": "重大のみ",
-  "Watch only": "監視のみ",
-  "Normal only": "通常のみ",
-  "ALL RISK": "すべてのリスク",
-  "All risk": "すべてのリスク",
-  "CRITICAL RISK": "重大リスク",
-  "Critical risk": "重大リスク",
-  "WATCH RISK": "監視リスク",
-  "Watch risk": "監視リスク",
-  "NORMAL RISK": "通常リスク",
-  "Normal risk": "通常リスク",
-  "Select a source risk filter to batch-revoke risky groups.": "ソースリスクフィルターを選択してリスクグループを一括取り消しします。",
-  "EXPORT RISK GROUPS CSV": "リスクグループ CSV をエクスポート",
-  "Export risk groups CSV": "リスクグループ CSV をエクスポート",
-  "DURABLE RECOVERY": "永続リカバリー",
-  "Durable recovery": "永続リカバリー",
-  "Inspect durable waiting-session compensation tasks, replay exhausted retries, and clear stale failures without leaving the console.":
-    "コンソールを離れずに永続的な待機セッション補償タスクを確認し、リトライを使い切ったタスクを再実行し、古い失敗をクリアします。",
-  "All task types": "すべてのタスクタイプ",
-  "Replay matching failed tasks": "一致する失敗タスクを再実行",
-  "Queuing replay...": "再実行をキューに追加中…",
-  "Resume is available for worlds that still need to move back toward running, including sleeping, failed, queued, and stopping states. Suspend is limited to worlds that are currently active. Retry is reserved for failed or in-flight lifecycle states.":
-    "再開はまだ実行状態に戻る必要があるワールド（スリープ、失敗、キュー済み、停止中の状態を含む）で利用できます。停止は現在アクティブなワールドのみ、リトライは失敗または進行中のライフサイクル状態のみに使用できます。",
-  Deployment: "デプロイ",
-  "Default region": "デフォルトリージョン",
-  "Default zone": "デフォルトゾーン",
-  "Managed provisioning": "マネージドプロビジョニング",
-  Snapshots: "スナップショット",
-  Yes: "はい",
-  No: "いいえ",
-  "Export risk sessions CSV": "リスクセッション CSV をエクスポート",
-  "Revoke matching risk groups": "一致するリスクグループを取り消し",
-  "Revoke group": "グループを取り消し",
-  "Revoke now": "今すぐ取り消し",
-  "Revoke selected": "選択したセッションを取り消し",
-  "Viewing sessions": "セッションを表示中",
-  "View sessions": "セッションを表示",
-  "No source groups match this filter.": "このフィルターに一致するソースグループはありません。",
-  "Select active sessions on this page to revoke them in one action.":
-    "このページのアクティブセッションを選択すると一括で取り消せます。",
-  "{count} active session(s) selected on this page.": "このページで {count} 件のアクティブセッションを選択しました。",
-  "Export filtered snapshot": "フィルタースナップショットをエクスポート",
-  "Export filtered CSV": "フィルター CSV をエクスポート",
-  "Export context groups snapshot": "コンテキストグループスナップショットをエクスポート",
-  "Export context groups CSV": "コンテキストグループ CSV をエクスポート",
-  "Open waiting sync permalink": "待機同期パーマリンクを開く",
-  "No visible contexts on this page yet.": "このページにはまだ表示中のコンテキストがありません。",
-  "Loading waiting sync tasks...": "待機同期タスクを読み込み中...",
-  "页 1 of 1": "1 / 1 ページ",
-  "Provider:": "プロバイダー:",
-  "Executor:": "実行モード:",
-  "Fail endpoint:": "失敗エンドポイント:",
-  "Reason:": "理由:",
-  "Escalated:": "格上げ済み:",
-  "Escalation reason:": "格上げ理由:",
-  "Not stale": "古くない",
-  Disabled: "無効",
-  "Queue totals reflect all jobs for this world, not just the recent 20 jobs below.":
-    "キューの合計はこのワールドのすべてのジョブを反映しており、下記の最近 20 件だけではありません。",
-  "Rotating...": "ローテーション中…",
-  "Rotate callback token": "コールバックトークンをローテーション",
-  "Phone unavailable": "電話番号なし",
-  "No tasks match this filter.": "このフィルターに一致するタスクはありません。",
-  "Previous groups": "前のグループ",
-  "Next groups": "次のグループ",
-  "Event view": "イベントビュー",
-  "Current snapshot": "現在のスナップショット",
-  "Session issued": "セッション発行",
-  "Session expired": "セッション期限切れ",
-  "Session revoked": "セッション取り消し",
-  "Unknown client": "不明なクライアント",
-  "Revoke session": "セッションを取り消し",
-  "Open user": "ユーザーを開く",
-  remaining: "残り",
-  expires: "期限",
-  available: "利用可能",
-};
-
-const cloudConsoleKoreanTextOverrides: Record<string, string> = {
-  Actions: "작업",
-  Access: "접근",
-  Action: "작업",
-  Active: "활성",
-  Admin: "관리자",
-  "Config key": "설정 키",
-  "Internal note shown next to the key.": "키 옆에 표시되는 내부 메모입니다.",
-  "Value (JSON)": "값 (JSON)",
-  "manual-ban": "수동 차단",
-  "Saving...": "저장 중…",
-  "Save world": "월드 저장",
-  "Refreshing...": "새로 고침 중…",
-  "Refresh status": "상태 새로 고침",
-  "Health message": "헬스 메시지",
-  Name: "이름",
-  "Provider instance": "프로바이더 인스턴스",
-  "Provider volume": "프로바이더 볼륨",
-  "Provider snapshot": "프로바이더 스냅샷",
-  "Private IP": "사설 IP",
-  "Public IP": "공용 IP",
-  Region: "리전",
-  Zone: "존",
-  Image: "이미지",
-  Flavor: "사양",
-  Disk: "디스크",
-  Bootstrapped: "부트스트랩 완료",
-  Created: "생성 시각",
-  Updated: "업데이트 시각",
-  "Provider message": "프로바이더 메시지",
-  "Deployment state": "배포 상태",
-  "Deployment mode": "배포 모드",
-  "Executor mode": "실행 모드",
-  "Remote host": "원격 호스트",
-  "Remote path": "원격 경로",
-  Project: "프로젝트",
-  Container: "컨테이너",
-  "Raw status": "원시 상태",
-  "No instance record exists yet. Provisioning will create one automatically.":
-    "아직 인스턴스 레코드가 없습니다. 다음 프로비저닝에서 자동으로 생성됩니다.",
-  "Provider-side deployment status observed from the current compute adapter.":
-    "현재 컴퓨트 어댑터에서 관찰한 프로바이더 측 배포 상태입니다.",
-  "Jobs show how this world moved through provision, resume, and suspend work.":
-    "작업은 이 월드가 프로비저닝, 재개, 일시 중지 등 라이프사이클을 거친 과정을 보여줍니다.",
-  "Plan code": "플랜 코드",
-  "Plan name": "플랜 이름",
-  "Duration (days)": "기간 (일)",
-  "Price (cents)": "가격 (센트)",
-  Currency: "통화",
-  "Sort order": "정렬 순서",
-  Description: "설명",
-  Trial: "체험",
-  "Publicly purchasable": "공개 구매 가능",
-  "e.g. trial": "예: trial",
-  "e.g. New user trial": "예: 신규 사용자 체험",
-  "Shown to users in the app subscription page.": "앱 구독 페이지에서 사용자에게 표시되는 문구입니다.",
-  Code: "코드",
-  "Admin:": "관리자:",
-  API: "API",
-  "API:": "API:",
-  "Artifact summary": "아티팩트 요약",
-  "Alert status": "알림 상태",
-  "Attention worlds": "주의 월드",
-  "attention:": "주의:",
-  "audit:": "감사:",
-  "active providers": "활성 프로바이더",
-  "Activity endpoint:": "활동 엔드포인트:",
-  "Approval and manual delivery queue shortcuts wired to the request projection filters.":
-    "승인 및 수동 인계 큐 바로가기가 요청 프로젝션 필터에 연결되어 있습니다.",
-  "Approval and projected-world shortcuts surfaced in the main dashboard header.":
-    "승인 및 예상 월드 바로가기가 기본 대시보드 헤더에 표시됩니다.",
-  "Accounts and expiry": "계정 및 만료",
-  "Aggregate sessions by issue IP and client under the current filters, then revoke an entire source in one action.":
-    "현재 필터 기준으로 발급 IP 및 클라이언트별 세션을 집계하고 소스 전체를 한 번에 취소할 수 있습니다.",
-  "Alerts upgraded because retry or stale-heartbeat thresholds were crossed.":
-    "재시도 또는 하트비트 부실 임계값을 초과하여 알림이 격상되었습니다.",
-  "Bootstrap endpoint:": "부트스트랩 엔드포인트:",
-  "Bootstrapped:": "부트스트랩 완료:",
-  "Callback token rotated.": "콜백 토큰이 교체되었습니다.",
-  "Callback token:": "콜백 토큰:",
-  "Clear context focus": "컨텍스트 포커스 해제",
-  "Clear matching failed tasks": "일치하는 실패 작업 정리",
-  "Clear now": "지금 정리",
-  "Clear receipts": "영수증 정리",
-  "Clear source focus": "소스 포커스 해제",
-  "Clear task highlight": "작업 하이라이트 해제",
-  "Clear world scope": "월드 범위 해제",
-  "Close review": "리뷰 닫기",
-  "Cloud platform:": "클라우드 플랫폼:",
-  "Configure role usage pricing, payees, contribution weights, and settlement ledgers.":
-    "역할 사용 가격, 수령인, 기여 가중치, 정산 원장을 설정합니다.",
-  "Configure trial, monthly, quarterly, yearly, and invite reward plans shown to the app.":
-    "앱에 표시되는 체험, 월간, 분기, 연간, 초대 보상 플랜을 설정합니다.",
-  "Compatibility view for the older approval workflow. Staff can still use it when manual intervention is needed.":
-    "기존 승인 워크플로의 호환 보기입니다. 수동 개입이 필요할 때 운영자가 계속 사용할 수 있습니다.",
-  "Context groups": "컨텍스트 그룹",
-  "Container:": "컨테이너:",
-  "Copy compose": "Compose 복사",
-  "Copy endpoints": "엔드포인트 복사",
-  "Copy env": "환경 변수 복사",
-  "Copy jobs permalink": "작업 퍼머링크 복사",
-  "Copy requests permalink": "요청 퍼머링크 복사",
-  "Copy review context": "리뷰 컨텍스트 복사",
-  "Copy sessions permalink": "세션 퍼머링크 복사",
-  "Copy task context": "작업 컨텍스트 복사",
-  "Copy token": "토큰 복사",
-  "Copy waiting sync permalink": "대기 동기화 퍼머링크 복사",
-  "Copy worlds permalink": "월드 퍼머링크 복사",
-  "Critical alerts": "심각 알림",
-  "Critical stale threshold:": "심각 하트비트 부실 임계값:",
-  "Critical threshold": "심각 임계값",
-  "Current alert severity after applying retry and stale-heartbeat thresholds.":
-    "재시도 및 하트비트 부실 임계값을 적용한 후의 현재 알림 심각도입니다.",
-  "Current endpoints": "현재 엔드포인트",
-  "Current rationale": "현재 근거",
-  "Current session": "현재 세션",
-  "Daily summary": "일별 요약",
-  "Default region:": "기본 리전:",
-  "Default zone:": "기본 존:",
-  Delayed: "지연",
-  Running: "실행 중",
-  Other: "기타",
-  "Lease expired": "리스 만료",
-  "{count} jobs": "{count}개 작업",
-  "Deployment mode:": "배포 모드:",
-  "Deployment state:": "배포 상태:",
-  "Deployment:": "배포:",
-  "Derived from session issue, expiry, and revoke events inside the focused source group under the current filters.":
-    "현재 필터에서 포커스된 소스 그룹 내의 세션 발급, 만료, 취소 이벤트로부터 생성됩니다.",
-  "Desired state": "기대 상태",
-  Direction: "방향",
-  "Download CSV": "CSV 다운로드",
-  "Download JSON": "JSON 다운로드",
-  "Export context CSV": "컨텍스트 CSV 내보내기",
-  "Export context snapshot": "컨텍스트 스냅샷 내보내기",
-  "Export focus CSV": "포커스 CSV 내보내기",
-  "Export focus snapshot": "포커스 스냅샷 내보내기",
-  "Export matching CSV": "일치 CSV 내보내기",
-  "Export matching snapshot": "일치 스냅샷 내보내기",
-  "Export risk CSV": "리스크 CSV 내보내기",
-  "Export risk snapshot": "리스크 스냅샷 내보내기",
-  "Export snapshot": "스냅샷 내보내기",
-  Escalated: "격상됨",
-  "Failure endpoint:": "실패 엔드포인트:",
-  "Focused source": "포커스된 소스",
-  "Focus path:": "포커스 경로:",
-  "Group risk": "그룹 리스크",
-  "Manage SaaS trial, invite, feature, copy, and public app URL configuration.":
-    "SaaS 체험, 초대, 기능, 문구 및 공개 앱 URL 설정을 관리합니다.",
-  "Health endpoint:": "헬스 엔드포인트:",
-  "Heartbeat endpoint:": "하트비트 엔드포인트:",
-  "Heartbeat stale": "하트비트 부실",
-  "Image:": "이미지:",
-  Instance: "인스턴스",
-  "Issued from": "발급 출처",
-  "Job status": "작업 상태",
-  "Last client": "최근 클라이언트",
-  "Manual Docker": "수동 Docker",
-  "Managed lifecycle": "관리형 라이프사이클",
-  "Matching sessions": "일치 세션",
-  "No requests match this filter.": "이 필터에 일치하는 요청이 없습니다.",
-  "No worlds match this filter.": "이 필터에 일치하는 월드가 없습니다.",
-  "No jobs match this filter.": "이 필터에 일치하는 작업이 없습니다.",
-  "No waiting sync tasks match this filter.": "이 필터에 일치하는 대기 동기화 작업이 없습니다.",
-  "No admin sessions match this filter.": "이 필터에 일치하는 관리 세션이 없습니다.",
-  "No active attention items. The fleet currently looks healthy.":
-    "활성 주의 항목이 없습니다. 현재 플릿은 정상으로 보입니다.",
-  "No pending or running jobs in the operator queue.": "운영자 큐에 보류 중이거나 실행 중인 작업이 없습니다.",
-  "No provider allocation data yet.": "아직 프로바이더 할당 데이터가 없습니다.",
-  "No recent failed jobs.": "최근 실패 작업이 없습니다.",
-  "No recent superseded jobs.": "최근 대체된 작업이 없습니다.",
-  "No request workflow items yet.": "아직 요청 워크플로 항목이 없습니다.",
-  Next: "다음",
-  "Ops note": "운영 메모",
-  Page: "페이지",
-  "Payees and ledgers": "수령인 및 원장",
-  "Page size": "페이지 크기",
-  "per page": "페이지당",
-  Previous: "이전",
-  "Pricing and access": "가격 및 접근",
-  "Project:": "프로젝트:",
-  "Projected world": "예상 월드",
-  "power:": "전원:",
-  "provider:": "프로바이더:",
-  "provider: all": "프로바이더: 전체",
-  "queue:": "큐:",
-  "queue: all": "큐: 전체",
-  "queue: running": "큐: 실행 중",
-  "queue: lease expired": "큐: 리스 만료",
-  "queue: delayed": "큐: 지연",
-  "Recent operation receipts": "최근 작업 영수증",
-  "Remote host:": "원격 호스트:",
-  "Remote path:": "원격 경로:",
-  "Replay now": "지금 재실행",
-  "Request Workflow": "요청 워크플로",
-  "Reset filters": "필터 초기화",
-  Result: "결과",
-  "Retry count": "재시도 횟수",
-  "Retry count:": "재시도 횟수:",
-  "Retry threshold:": "재시도 임계값:",
-  "Review live admin sessions, inspect where they were issued from, filter by revocation path, and page through longer audit history.":
-    "실시간 관리 세션을 확인하고 발급 출처를 조사하며 취소 경로로 필터링하고 더 긴 감사 기록을 페이지로 탐색합니다.",
-  "Review cloud account status, subscription expiry, inviter attribution, and world linkage.":
-    "클라우드 계정 상태, 구독 만료, 초대자 귀속, 월드 연결을 확인합니다.",
-  "Review reward redemptions, device and IP signals, and manual reversals.":
-    "보상 사용, 기기 및 IP 신호, 수동 취소를 확인합니다.",
-  "Review requests (": "요청 검토 (",
-  "Rewards and risk": "보상 및 리스크",
-  Reviewing: "검토 중",
-  Revocation: "취소",
-  "Revocation reason": "취소 사유",
-  "Revoked by session": "세션에 의한 취소",
-  "Risk timeline": "리스크 타임라인",
-  "Risk timeline data is not ready for export yet.": "리스크 타임라인 데이터를 아직 내보낼 수 없습니다.",
-  Runbook: "런북",
-  "Running now": "실행 중",
-  "Runtime env overlay": "런타임 환경 오버레이",
-  "Runtime is not checking in within the configured stale window.":
-    "런타임이 설정된 부실 윈도우 내에 체크인하지 않았습니다.",
-  "Runtime observation": "런타임 관찰",
-  Scope: "범위",
-  "Search requests": "요청 검색",
-  "Search worlds": "월드 검색",
-  "Select all active admin sessions": "모든 활성 관리 세션 선택",
-  "Select provider": "프로바이더 선택",
-  "Session context": "세션 컨텍스트",
-  Showing: "표시",
-  "Showing the latest": "최신 표시 중",
-  "Snapshots:": "스냅샷:",
-  "Sort by": "정렬 기준",
-  "sort by:": "정렬:",
-  "direction:": "방향:",
-  "Source context": "소스 컨텍스트",
-  "Source direction": "소스 정렬 방향",
-  "Source groups": "소스 그룹",
-  "Source page size": "소스 페이지 크기",
-  "Source risk": "소스 리스크",
-  "Source sort": "소스 정렬",
-  "Stale heartbeat seconds:": "하트비트 부실 초:",
-  "status:": "상태:",
-  "health:": "헬스:",
-  "superseded by:": "대체 출처:",
-  "Superseded Queue": "대체된 큐",
-  "Suggested admin:": "권장 관리 주소:",
-  "Suggested API:": "권장 API:",
-  "Summary counts reflect all jobs matching the current filters, not just this page.":
-    "요약 카운트는 현재 필터에 일치하는 모든 작업을 반영하며 이 페이지만 보여주는 것이 아닙니다.",
-  "Synced with": "동기화 대상",
-  "Target values:": "목표 값:",
-  Targets: "타겟",
-  Task: "작업",
-  "Trial and copy": "체험 및 문구",
-  "Task highlighted": "작업 하이라이트됨",
-  "Task permalink focus": "작업 퍼머링크 포커스",
-  "Task types:": "작업 타입:",
-  "The world will move toward sleeping state and active sessions may need to reconnect after it wakes again.":
-    "월드가 슬립 상태로 전환되며 다시 깨어난 후 활성 세션이 재연결해야 할 수 있습니다.",
-  "These sessions reflect the latest event captured in this summary.":
-    "이러한 세션은 이 요약에서 캡처한 최신 이벤트를 반영합니다.",
-  "This panel folds together runtime heartbeat freshness, provider-observed drift, and queued recovery jobs.":
-    "이 패널은 런타임 하트비트 신선도, 프로바이더 관찰 드리프트, 큐의 복구 작업을 함께 보여줍니다.",
-  "This provider key is not in the current catalog yet.": "이 프로바이더 키는 아직 현재 카탈로그에 없습니다.",
-  "This task is no longer visible in the current result set, but recent receipts still match this task permalink.":
-    "이 작업은 현재 결과 집합에 더 이상 표시되지 않지만 최근 영수증은 여전히 이 작업 퍼머링크와 일치합니다.",
-  "This will queue a new recovery action and clear the current failure state for the world.":
-    "새 복구 작업을 큐에 추가하고 이 월드의 현재 실패 상태를 정리합니다.",
-  "Timeline audit detail": "타임라인 감사 상세",
-  "Timeline focus": "타임라인 포커스",
-  Timing: "시각",
-  "Top attention items": "주요 주의 항목",
-  "Total fleet:": "전체 플릿:",
-  "Total results": "전체 결과 수",
-  "type:": "타입:",
-  Unassigned: "미할당",
-  "Use this env overlay when deploying the user's dedicated world runtime.":
-    "사용자 전용 월드 런타임을 배포할 때 이 환경 변수 오버레이를 사용하세요.",
-  "View in sessions list": "세션 목록에서 보기",
-  "Viewing source group": "소스 그룹 보기 중",
-  "Visible contexts from the current page. Use these cards to focus a single recovery path before replaying or clearing failed tasks.":
-    "현재 페이지에 표시되는 컨텍스트입니다. 실패 작업을 재실행하거나 정리하기 전에 이 카드로 단일 복구 경로에 포커스할 수 있습니다.",
-  "Visible failed": "표시된 실패",
-  "Visible pending": "표시된 보류",
-  "Visible rows": "표시된 행",
-  "Visible running": "표시된 실행 중",
-  "visible task(s) · latest update": "개 표시 작업 · 최근 업데이트",
-  "visible task(s) from": "개 표시 작업, 출처",
-  "Volume:": "볼륨:",
-  "Waiting sync task clear was skipped.": "대기 동기화 작업 정리가 건너뛰어졌습니다.",
-  "Waiting sync task cleared.": "대기 동기화 작업을 정리했습니다.",
-  "Waiting sync task replay queued.": "대기 동기화 작업 재실행을 큐에 추가했습니다.",
-  "Waiting sync task replay was skipped.": "대기 동기화 작업 재실행이 건너뛰어졌습니다.",
-  "Watch threshold": "관찰 임계값",
-  "Weekly summary": "주별 요약",
-  "Work from the instance point of view: provider placement, power state, heartbeat freshness, and quick lifecycle actions.":
-    "인스턴스 관점의 작업: 프로바이더 배치, 전원 상태, 하트비트 신선도, 빠른 라이프사이클 동작.",
-  "World detail:": "월드 상세:",
-  "World drift summary": "월드 드리프트 요약",
-  "World failed jobs": "월드 실패 작업",
-  "World operator jobs": "월드 운영 작업",
-  "World request saved.": "월드 요청을 저장했습니다.",
-  "World scope": "월드 범위",
-  "World settings saved.": "월드 설정을 저장했습니다.",
-  "World superseded jobs": "월드 대체 작업",
-  "Worlds already in critical state, including failed and escalated alerts.":
-    "이미 심각 상태인 월드 (실패 및 격상된 알림 포함).",
-  "Worlds that already have active `resume` or `provision` work in flight.":
-    "이미 활성 `resume` 또는 `provision` 작업이 진행 중인 월드.",
-  "Worlds that currently need operator attention.": "현재 운영자 주의가 필요한 월드.",
-  "Zone:": "존:",
-  "Quick view of world availability, queued recovery, and the most urgent runtime drift signals across the cloud fleet.":
-    "클라우드 플릿 전체의 월드 가용성, 큐의 복구, 가장 긴급한 런타임 드리프트 신호를 빠르게 봅니다.",
-  "Inspect jobs": "작업 확인",
-  "PROJECTED:": "예상:",
-  "DESIRED:": "기대:",
-  "Projected:": "예상:",
-  "Desired:": "기대:",
-  Creating: "생성 중",
-  CREATING: "생성 중",
-  Bootstrapping: "부트스트랩 중",
-  BOOTSTRAPPING: "부트스트랩 중",
-  Cancelled: "취소됨",
-  QUEUED: "큐에 등록됨",
-  STARTING: "시작 중",
-  READY: "준비 완료",
-  SLEEPING: "슬립",
-  FAILED: "실패",
-  DISABLED: "비활성",
-  ALL: "전체",
-  all: "전체",
-  active: "활성",
-  revoked: "취소됨",
-  total: "건",
-  expired: "만료됨",
-  "refresh reuse": "리프레시 토큰 재사용",
-  pending: "보류",
-  running: "실행 중",
-  succeeded: "성공",
-  failed: "실패",
-  cancelled: "취소됨",
-  queued: "큐에 등록됨",
-  creating: "생성 중",
-  bootstrapping: "부트스트랩 중",
-  starting: "시작 중",
-  ready: "준비 완료",
-  sleeping: "슬립",
-  disabled: "비활성",
-  absent: "인스턴스 없음",
-  provisioning: "프로비저닝 중",
-  stopped: "중지됨",
-  stopping: "중지 중",
-  error: "오류",
-  healthy: "정상",
-  critical: "심각",
-  warning: "경고",
-  info: "정보",
-  unhealthy: "비정상",
-  unknown: "알 수 없음",
-  normal: "정상",
-  watch: "관찰",
-  provision: "프로비저닝",
-  resume: "재개",
-  suspend: "일시 중지",
-  reconcile: "정합",
-  superseded: "대체됨",
-  refresh_world: "월드 새로 고침",
-  refresh_phone: "전화번호 새로 고침",
-  invalidate_phone: "전화번호 무효화",
-  updatedAt: "업데이트 시각",
-  createdAt: "생성 시각",
-  availableAt: "사용 가능 시각",
-  startedAt: "시작 시각",
-  finishedAt: "완료 시각",
-  expiresAt: "리프레시 만료",
-  lastUsedAt: "최근 사용",
-  revokedAt: "취소 시각",
-  activeSessions: "활성 세션",
-  totalSessions: "전체 세션",
-  latestLastUsedAt: "최근 사용",
-  latestCreatedAt: "최근 생성",
-  latestRevokedAt: "최근 취소",
-  desc: "내림차순",
-  asc: "오름차순",
-  current: "현재",
-  logout: "로그아웃",
-  "manual-revocation": "수동 취소",
-  "refresh-token-reuse": "리프레시 토큰 재사용",
-  running_now: "실행 중",
-  lease_expired: "리스 만료",
-  "Warning:": "경고:",
-  "Heartbeat stale:": "하트비트 부실:",
-  "Failed:": "실패:",
-  "Instance pool": "인스턴스 풀",
-  "Instance fleet": "인스턴스 플릿",
-  worlds: "개 월드",
-  WORLDS: "개 월드",
-  provider: "프로바이더",
-  PROVIDER: "프로바이더",
-  Provider: "프로바이더",
-  "Provider drift": "프로바이더 드리프트",
-  "Provider spread": "프로바이더 분포",
-  "Open requests (": "요청 열기 (",
-  "Open failed jobs (": "실패 작업 열기 (",
-  "Open superseded jobs (": "대체 작업 열기 (",
-  "Open superseded queue (": "대체 큐 열기 (",
-  "Open world fleet": "월드 플릿 열기",
-  "Loading request workflow...": "요청 워크플로 로드 중...",
-  "Operator Queue": "운영자 큐",
-  "Pending and running lifecycle work across the instance fleet.":
-    "인스턴스 플릿 전체의 보류 및 실행 중 라이프사이클 작업입니다.",
-  "Loading lifecycle jobs...": "라이프사이클 작업 로드 중...",
-  "Recent Failures": "최근 실패",
-  "Latest failed lifecycle work that may need manual recovery.":
-    "수동 복구가 필요할 수 있는 최근 실패 라이프사이클 작업입니다.",
-  "Latest lifecycle jobs that were replaced by newer work.": "새 작업으로 대체된 최근 라이프사이클 작업입니다.",
-  "Loading superseded lifecycle jobs...": "대체 라이프사이클 작업 로드 중...",
-  "Most urgent worlds needing manual inspection or follow-up.": "수동 검사 또는 후속 작업이 가장 긴급한 월드입니다.",
-  "Loading fleet dashboard...": "플릿 대시보드 로드 중...",
-  "Loading requests...": "요청 로드 중...",
-  "Loading worlds...": "월드 로드 중...",
-  "Loading drift summary...": "드리프트 요약 로드 중...",
-  "Loading world drift summary...": "월드 드리프트 요약 로드 중...",
-  "Loading instances...": "인스턴스 로드 중...",
-  "Loading jobs...": "작업 로드 중...",
-  "World name": "월드 이름",
-  Lease: "리스",
-  "Provider reports power state that disagrees with desired world state.":
-    "프로바이더가 보고한 전원 상태가 월드 기대 상태와 일치하지 않습니다.",
-  "Each phone owns exactly one world. New users provision a fresh instance, while returning users wake their previous one.":
-    "각 전화번호는 정확히 하나의 월드를 소유합니다. 신규 사용자는 새 인스턴스를 프로비저닝하고 재방문 사용자는 이전 인스턴스를 깨웁니다.",
-  "NO INSTANCE": "인스턴스 없음",
-  "No instance": "인스턴스 없음",
-  "NEEDS ACTION": "조치 필요",
-  "Needs action": "조치 필요",
-  "power: all": "전원: 전체",
-  "power: absent": "전원: 인스턴스 없음",
-  "power: provisioning": "전원: 프로비저닝 중",
-  "power: running": "전원: 실행 중",
-  "power: starting": "전원: 시작 중",
-  "power: stopped": "전원: 중지됨",
-  "power: stopping": "전원: 중지 중",
-  "power: error": "전원: 오류",
-  "attention: all": "주의: 전체",
-  "attention: healthy": "주의: 정상",
-  "attention: critical": "주의: 심각",
-  "attention: warning": "주의: 경고",
-  "attention: info": "주의: 정보",
-  "health: all": "헬스: 전체",
-  "health: healthy": "헬스: 정상",
-  "health: unhealthy": "헬스: 비정상",
-  "health: unknown": "헬스: 알 수 없음",
-  "Inspect provisioning, resume, suspend, and reconcile work across the managed world fleet.":
-    "관리형 월드 플릿 전체의 프로비저닝, 재개, 일시 중지, 정합 작업을 확인합니다.",
-  "status: all": "상태: 전체",
-  "status: pending": "상태: 보류",
-  "status: running": "상태: 실행 중",
-  "status: succeeded": "상태: 성공",
-  "status: failed": "상태: 실패",
-  "status: cancelled": "상태: 취소됨",
-  "type: all": "타입: 전체",
-  "type: provision": "타입: 프로비저닝",
-  "type: resume": "타입: 재개",
-  "type: suspend": "타입: 일시 중지",
-  "type: reconcile": "타입: 정합",
-  "audit: all": "감사: 전체",
-  "audit: superseded": "감사: 대체됨",
-  "superseded by: all": "대체 출처: 전체",
-  "superseded by: provision": "대체 출처: 프로비저닝",
-  "superseded by: resume": "대체 출처: 재개",
-  "superseded by: suspend": "대체 출처: 일시 중지",
-  "superseded by: reconcile": "대체 출처: 정합",
-  "sort by: updatedAt": "정렬: 업데이트 시각",
-  "sort by: createdAt": "정렬: 생성 시각",
-  "sort by: availableAt": "정렬: 사용 가능 시각",
-  "sort by: startedAt": "정렬: 시작 시각",
-  "sort by: finishedAt": "정렬: 완료 시각",
-  "direction: desc": "방향: 내림차순",
-  "direction: asc": "방향: 오름차순",
-  "page size: 20": "페이지 크기: 20",
-  "page size: 50": "페이지 크기: 50",
-  "page size: 100": "페이지 크기: 100",
-  "Previous page": "이전 페이지",
-  "Next page": "다음 페이지",
-  "Showing ": "표시 중 ",
-  " of ": " / ",
-  " jobs.": "개 작업.",
-  "SHOWING 0 SESSIONS": "0개 세션 표시 중",
-  "Showing 0 sessions": "0개 세션 표시 중",
-  "SHOWING 0 GROUPS": "0개 그룹 표시 중",
-  "Showing 0 groups": "0개 그룹 표시 중",
-  "QUICK VIEWS": "빠른 보기",
-  "Quick views": "빠른 보기",
-  "RECENTLY REVOKED": "최근 취소",
-  "Recently revoked": "최근 취소",
-  "EXPIRING SOON": "곧 만료",
-  "Expiring soon": "곧 만료",
-  "Loading source groups...": "소스 그룹 로드 중...",
-  "Loading admin sessions...": "관리 세션 로드 중...",
-  "REVOKE ALL MATCHING": "일치 항목 모두 취소",
-  "Revoke all matching": "일치 항목 모두 취소",
-  "Manual revoke": "수동 취소",
-  "Refresh reuse": "리프레시 재사용",
-  "Refresh expiry": "리프레시 만료",
-  "Active sessions": "활성 세션",
-  "Total sessions": "전체 세션",
-  "Latest used": "최근 사용",
-  "Latest created": "최근 생성",
-  "Latest revoked": "최근 취소",
-  "All risk levels": "모든 리스크 레벨",
-  "Critical only": "심각만",
-  "Watch only": "관찰만",
-  "Normal only": "정상만",
-  "ALL RISK": "모든 리스크",
-  "All risk": "모든 리스크",
-  "CRITICAL RISK": "심각 리스크",
-  "Critical risk": "심각 리스크",
-  "WATCH RISK": "관찰 리스크",
-  "Watch risk": "관찰 리스크",
-  "NORMAL RISK": "정상 리스크",
-  "Normal risk": "정상 리스크",
-  "Select a source risk filter to batch-revoke risky groups.":
-    "소스 리스크 필터를 선택하여 리스크 그룹을 일괄 취소할 수 있습니다.",
-  "EXPORT RISK GROUPS CSV": "리스크 그룹 CSV 내보내기",
-  "Export risk groups CSV": "리스크 그룹 CSV 내보내기",
-  "DURABLE RECOVERY": "내구성 복구",
-  "Durable recovery": "내구성 복구",
-  "Inspect durable waiting-session compensation tasks, replay exhausted retries, and clear stale failures without leaving the console.":
-    "콘솔을 떠나지 않고 내구성 대기 세션 보정 작업을 확인하고, 재시도가 소진된 작업을 재실행하며, 부실 실패를 정리합니다.",
-  "All task types": "모든 작업 타입",
-  "Replay matching failed tasks": "일치하는 실패 작업 재실행",
-  "Queuing replay...": "재실행 큐에 추가 중…",
-  "Resume is available for worlds that still need to move back toward running, including sleeping, failed, queued, and stopping states. Suspend is limited to worlds that are currently active. Retry is reserved for failed or in-flight lifecycle states.":
-    "재개는 다시 실행 상태로 돌아가야 하는 월드 (슬립, 실패, 큐 등록, 중지 중 상태 포함)에서 사용할 수 있습니다. 일시 중지는 현재 활성인 월드에만 적용되며, 재시도는 실패 또는 진행 중인 라이프사이클 상태에만 사용할 수 있습니다.",
-  Deployment: "배포",
-  "Default region": "기본 리전",
-  "Default zone": "기본 존",
-  "Managed provisioning": "관리형 프로비저닝",
-  Snapshots: "스냅샷",
-  Yes: "예",
-  No: "아니오",
-  "Export risk sessions CSV": "리스크 세션 CSV 내보내기",
-  "Revoke matching risk groups": "일치하는 리스크 그룹 취소",
-  "Revoke group": "그룹 취소",
-  "Revoke now": "지금 취소",
-  "Revoke selected": "선택한 세션 취소",
-  "Viewing sessions": "세션 보기 중",
-  "View sessions": "세션 보기",
-  "No source groups match this filter.": "이 필터에 일치하는 소스 그룹이 없습니다.",
-  "Select active sessions on this page to revoke them in one action.":
-    "이 페이지의 활성 세션을 선택하면 한 번에 취소할 수 있습니다.",
-  "{count} active session(s) selected on this page.": "이 페이지에서 {count}개 활성 세션을 선택했습니다.",
-  "Export filtered snapshot": "필터 스냅샷 내보내기",
-  "Export filtered CSV": "필터 CSV 내보내기",
-  "Export context groups snapshot": "컨텍스트 그룹 스냅샷 내보내기",
-  "Export context groups CSV": "컨텍스트 그룹 CSV 내보내기",
-  "Open waiting sync permalink": "대기 동기화 퍼머링크 열기",
-  "No visible contexts on this page yet.": "이 페이지에 아직 표시된 컨텍스트가 없습니다.",
-  "Loading waiting sync tasks...": "대기 동기화 작업 로드 중...",
-  "页 1 of 1": "1 / 1 페이지",
-  "Provider:": "프로바이더:",
-  "Executor:": "실행 모드:",
-  "Fail endpoint:": "실패 엔드포인트:",
-  "Reason:": "사유:",
-  "Escalated:": "격상됨:",
-  "Escalation reason:": "격상 사유:",
-  "Not stale": "부실 아님",
-  Disabled: "비활성",
-  "Queue totals reflect all jobs for this world, not just the recent 20 jobs below.":
-    "큐 총계는 이 월드의 모든 작업을 반영하며 아래 최근 20개 작업만이 아닙니다.",
-  "Rotating...": "교체 중…",
-  "Rotate callback token": "콜백 토큰 교체",
-  "Phone unavailable": "전화번호 없음",
-  "No tasks match this filter.": "이 필터에 일치하는 작업이 없습니다.",
-  "Previous groups": "이전 그룹",
-  "Next groups": "다음 그룹",
-  "Event view": "이벤트 보기",
-  "Current snapshot": "현재 스냅샷",
-  "Session issued": "세션 발급",
-  "Session expired": "세션 만료",
-  "Session revoked": "세션 취소",
-  "Unknown client": "알 수 없는 클라이언트",
-  "Revoke session": "세션 취소",
-  "Open user": "사용자 열기",
-  remaining: "남음",
-  expires: "만료",
-  available: "사용 가능",
-};
-
-const surfaceDictionaries: Partial<Record<I18nAppSurface, LocaleDictionary>> = {
-  "cloud-console": cloudConsoleText,
-};
+export function registerSurfaceTextDictionaryProvider(
+  surface: I18nAppSurface,
+  provider: DictionaryProvider,
+) {
+  dictionaryProviders.set(surface, provider);
+}
 
 export function getSurfaceTextDictionary(
   surface: I18nAppSurface,
   locale: SupportedLocale,
-) {
-  const dictionary = new Map(
-    Object.entries(surfaceDictionaries[surface]?.[locale] ?? {}),
-  );
-  if (surface === "cloud-console") {
-    const overrides =
-      locale === "zh-CN"
-        ? cloudConsoleChineseTextOverrides
-        : locale === "ja-JP"
-          ? cloudConsoleJapaneseTextOverrides
-          : locale === "ko-KR"
-            ? cloudConsoleKoreanTextOverrides
-            : null;
-    if (overrides) {
-      Object.entries(overrides).forEach(([key, value]) => {
-        dictionary.set(key, value);
-      });
-    }
+): Map<string, string> {
+  const provider = dictionaryProviders.get(surface);
+  if (!provider) {
+    return new Map();
   }
-  return dictionary;
+  return provider(locale);
 }
